Get data and return a data frame

The primary function in the rnassqs package, nassqs makes a HTTP GET request to the USDA-NASS Quick Stats API and returns the data parsed as a data.frame, plain text, or list. Various other functions make use of nassqs to make specific queries. For a data request the Quick Stats API returns JSON that when parsed to a data.frame contains 39 columns and a varying number of rows depending on the query. Unfortunately there is not a way to restrict the number of columns.

nassqs(..., as = c("data.frame", "text", "list"))

Arguments

...
either a named list of parameters or a series of parameters to form the query

as
whether to return a data.frame, list, or text string nassqs_GET()

Value

a data frame, list, or text string of requested data.

See also

Examples

  # Get corn yields in Virginia in 2012
  params <- list(commodity_name = "CORN",
                 year = 2012,
                 agg_level_desc = "COUNTY",
                 state_alpha = "VA",
                 statisticcat_desc = "YIELD")
  yields <- nassqs(params)

#>   |                                                                              |                                                                      |   0%  |                                                                              |=                                                                     |   2%  |                                                                              |====                                                                  |   5%  |                                                                              |=======                                                               |  10%  |                                                                              |===========                                                           |  15%  |                                                                              |==============                                                        |  20%  |                                                                              |==================                                                    |  26%  |                                                                              |===================                                                   |  26%  |                                                                              |======================                                                |  32%  |                                                                              |==========================                                            |  37%  |                                                                              |==========================                                            |  38%  |                                                                              |==============================                                        |  43%  |                                                                              |===============================                                       |  44%  |                                                                              |==================================                                    |  49%  |                                                                              |======================================                                |  54%  |                                                                              |=========================================                             |  59%  |                                                                              |=============================================                         |  64%  |                                                                              |=================================================                     |  69%  |                                                                              |=================================================                     |  70%  |                                                                              |=====================================================                 |  75%  |                                                                              |========================================================              |  81%  |                                                                              |=========================================================             |  81%  |                                                                              |=============================================================         |  87%  |                                                                              |================================================================      |  92%  |                                                                              |=================================================================     |  92%  |                                                                              |=================================================================     |  93%  |                                                                              |====================================================================  |  98%  |                                                                              |===================================================================== |  98%  |                                                                              |===================================================================== |  99%  |                                                                              |======================================================================| 100%

  head(yields)

#>    group_desc year region_desc      prodn_practice_desc unit_desc  class_desc
#> 1 FIELD CROPS 2012             ALL PRODUCTION PRACTICES BU / ACRE ALL CLASSES
#> 2 FIELD CROPS 2012             ALL PRODUCTION PRACTICES BU / ACRE ALL CLASSES
#> 3 FIELD CROPS 2012             ALL PRODUCTION PRACTICES BU / ACRE ALL CLASSES
#> 4 FIELD CROPS 2012             ALL PRODUCTION PRACTICES BU / ACRE ALL CLASSES
#> 5 FIELD CROPS 2012             ALL PRODUCTION PRACTICES BU / ACRE ALL CLASSES
#> 6 FIELD CROPS 2012             ALL PRODUCTION PRACTICES BU / ACRE ALL CLASSES
#>   watershed_code agg_level_desc sector_desc country_code end_code
#> 1       00000000         COUNTY       CROPS         9000       00
#> 2       00000000         COUNTY       CROPS         9000       00
#> 3       00000000         COUNTY       CROPS         9000       00
#> 4       00000000         COUNTY       CROPS         9000       00
#> 5       00000000         COUNTY       CROPS         9000       00
#> 6       00000000         COUNTY       CROPS         9000       00
#>   congr_district_code                            short_desc week_ending
#> 1                     BARLEY - YIELD, MEASURED IN BU / ACRE            
#> 2                     BARLEY - YIELD, MEASURED IN BU / ACRE            
#> 3                     BARLEY - YIELD, MEASURED IN BU / ACRE            
#> 4                     BARLEY - YIELD, MEASURED IN BU / ACRE            
#> 5                     BARLEY - YIELD, MEASURED IN BU / ACRE            
#> 6                     BARLEY - YIELD, MEASURED IN BU / ACRE            
#>   state_alpha source_desc domain_desc freq_desc reference_period_desc
#> 1          VA      SURVEY       TOTAL    ANNUAL                  YEAR
#> 2          VA      SURVEY       TOTAL    ANNUAL                  YEAR
#> 3          VA      SURVEY       TOTAL    ANNUAL                  YEAR
#> 4          VA      SURVEY       TOTAL    ANNUAL                  YEAR
#> 5          VA      SURVEY       TOTAL    ANNUAL                  YEAR
#> 6          VA      SURVEY       TOTAL    ANNUAL                  YEAR
#>    country_name CV (%) zip_5 state_fips_code state_name
#> 1 UNITED STATES                           51   VIRGINIA
#> 2 UNITED STATES                           51   VIRGINIA
#> 3 UNITED STATES                           51   VIRGINIA
#> 4 UNITED STATES                           51   VIRGINIA
#> 5 UNITED STATES                           51   VIRGINIA
#> 6 UNITED STATES                           51   VIRGINIA
#>                    location_desc           load_time statisticcat_desc
#> 1   VIRGINIA, NORTHERN, CULPEPER 2015-06-02 15:00:27             YIELD
#> 2   VIRGINIA, NORTHERN, FAUQUIER 2015-06-02 15:00:27             YIELD
#> 3    VIRGINIA, NORTHERN, MADISON 2015-06-02 15:00:27             YIELD
#> 4       VIRGINIA, NORTHERN, PAGE 2015-06-02 15:00:27             YIELD
#> 5 VIRGINIA, NORTHERN, ROCKINGHAM 2015-06-02 15:00:27             YIELD
#> 6 VIRGINIA, NORTHERN, SHENANDOAH 2015-06-02 15:00:27             YIELD
#>   county_code state_ansi domaincat_desc        util_practice_desc county_ansi
#> 1         047         51  NOT SPECIFIED ALL UTILIZATION PRACTICES         047
#> 2         061         51  NOT SPECIFIED ALL UTILIZATION PRACTICES         061
#> 3         113         51  NOT SPECIFIED ALL UTILIZATION PRACTICES         113
#> 4         139         51  NOT SPECIFIED ALL UTILIZATION PRACTICES         139
#> 5         165         51  NOT SPECIFIED ALL UTILIZATION PRACTICES         165
#> 6         171         51  NOT SPECIFIED ALL UTILIZATION PRACTICES         171
#>   asd_desc county_name watershed_desc begin_code Value asd_code commodity_desc
#> 1 NORTHERN    CULPEPER                        00  87.1       20         BARLEY
#> 2 NORTHERN    FAUQUIER                        00  68.1       20         BARLEY
#> 3 NORTHERN     MADISON                        00  71.8       20         BARLEY
#> 4 NORTHERN        PAGE                        00    67       20         BARLEY
#> 5 NORTHERN  ROCKINGHAM                        00    76       20         BARLEY
#> 6 NORTHERN  SHENANDOAH                        00    69       20         BARLEY