Get data and return a data frame

The primary function in the rnassqs package, nassqs makes a HTTP GET request to the USDA-NASS Quick Stats API and returns the data parsed as a data.frame, plain text, or list. Various other functions make use of nassqs to make specific queries. For a data request the Quick Stats API returns JSON that when parsed to a data.frame contains 39 columns and a varying number of rows depending on the query. Unfortunately there is not a way to restrict the number of columns.

nassqs(..., as = c("data.frame", "text", "list"))

Arguments

...
either a named list of parameters or a series of parameters to form the query

as
whether to return a data.frame, list, or text string nassqs_GET()

Value

a data frame, list, or text string of requested data.

See also

Examples

  # Get corn yields in Virginia in 2012
  params <- list(commodity_name = "CORN",
                 year = 2012,
                 agg_level_desc = "COUNTY",
                 state_alpha = "VA",
                 statisticcat_desc = "YIELD")
  yields <- nassqs(params)
  head(yields)

#>        prodn_practice_desc source_desc reference_period_desc  country_name
#> 1 ALL PRODUCTION PRACTICES      SURVEY                  YEAR UNITED STATES
#> 2 ALL PRODUCTION PRACTICES      SURVEY                  YEAR UNITED STATES
#> 3 ALL PRODUCTION PRACTICES      SURVEY                  YEAR UNITED STATES
#> 4 ALL PRODUCTION PRACTICES      SURVEY                  YEAR UNITED STATES
#> 5 ALL PRODUCTION PRACTICES      SURVEY                  YEAR UNITED STATES
#> 6 ALL PRODUCTION PRACTICES      SURVEY                  YEAR UNITED STATES
#>   watershed_desc agg_level_desc region_desc begin_code year domain_desc
#> 1                        COUNTY                     00 2012       TOTAL
#> 2                        COUNTY                     00 2012       TOTAL
#> 3                        COUNTY                     00 2012       TOTAL
#> 4                        COUNTY                     00 2012       TOTAL
#> 5                        COUNTY                     00 2012       TOTAL
#> 6                        COUNTY                     00 2012       TOTAL
#>   county_name end_code state_fips_code        util_practice_desc zip_5
#> 1    CULPEPER       00              51 ALL UTILIZATION PRACTICES      
#> 2    FAUQUIER       00              51 ALL UTILIZATION PRACTICES      
#> 3     MADISON       00              51 ALL UTILIZATION PRACTICES      
#> 4        PAGE       00              51 ALL UTILIZATION PRACTICES      
#> 5  ROCKINGHAM       00              51 ALL UTILIZATION PRACTICES      
#> 6  SHENANDOAH       00              51 ALL UTILIZATION PRACTICES      
#>   freq_desc domaincat_desc unit_desc                            short_desc
#> 1    ANNUAL  NOT SPECIFIED BU / ACRE BARLEY - YIELD, MEASURED IN BU / ACRE
#> 2    ANNUAL  NOT SPECIFIED BU / ACRE BARLEY - YIELD, MEASURED IN BU / ACRE
#> 3    ANNUAL  NOT SPECIFIED BU / ACRE BARLEY - YIELD, MEASURED IN BU / ACRE
#> 4    ANNUAL  NOT SPECIFIED BU / ACRE BARLEY - YIELD, MEASURED IN BU / ACRE
#> 5    ANNUAL  NOT SPECIFIED BU / ACRE BARLEY - YIELD, MEASURED IN BU / ACRE
#> 6    ANNUAL  NOT SPECIFIED BU / ACRE BARLEY - YIELD, MEASURED IN BU / ACRE
#>   country_code sector_desc  group_desc asd_code commodity_desc watershed_code
#> 1         9000       CROPS FIELD CROPS       20         BARLEY       00000000
#> 2         9000       CROPS FIELD CROPS       20         BARLEY       00000000
#> 3         9000       CROPS FIELD CROPS       20         BARLEY       00000000
#> 4         9000       CROPS FIELD CROPS       20         BARLEY       00000000
#> 5         9000       CROPS FIELD CROPS       20         BARLEY       00000000
#> 6         9000       CROPS FIELD CROPS       20         BARLEY       00000000
#>   asd_desc statisticcat_desc state_ansi state_alpha state_name county_code
#> 1 NORTHERN             YIELD         51          VA   VIRGINIA         047
#> 2 NORTHERN             YIELD         51          VA   VIRGINIA         061
#> 3 NORTHERN             YIELD         51          VA   VIRGINIA         113
#> 4 NORTHERN             YIELD         51          VA   VIRGINIA         139
#> 5 NORTHERN             YIELD         51          VA   VIRGINIA         165
#> 6 NORTHERN             YIELD         51          VA   VIRGINIA         171
#>   CV (%)  class_desc week_ending county_ansi                  location_desc
#> 1        ALL CLASSES                     047   VIRGINIA, NORTHERN, CULPEPER
#> 2        ALL CLASSES                     061   VIRGINIA, NORTHERN, FAUQUIER
#> 3        ALL CLASSES                     113    VIRGINIA, NORTHERN, MADISON
#> 4        ALL CLASSES                     139       VIRGINIA, NORTHERN, PAGE
#> 5        ALL CLASSES                     165 VIRGINIA, NORTHERN, ROCKINGHAM
#> 6        ALL CLASSES                     171 VIRGINIA, NORTHERN, SHENANDOAH
#>   Value congr_district_code           load_time
#> 1  87.1                     2015-06-02 15:00:27
#> 2  68.1                     2015-06-02 15:00:27
#> 3  71.8                     2015-06-02 15:00:27
#> 4    67                     2015-06-02 15:00:27
#> 5    76                     2015-06-02 15:00:27
#> 6    69                     2015-06-02 15:00:27