Get data and return a data frame

The primary function in the rnassqs package, nassqs makes a HTTP GET request to the USDA-NASS Quick Stats API and returns the data parsed as a data.frame, plain text, or list. Various other functions make use of nassqs to make specific queries. For a data request the Quick Stats API returns JSON that when parsed to a data.frame contains 39 columns and a varying number of rows depending on the query. Unfortunately there is not a way to restrict the number of columns.

nassqs(..., as = c("data.frame", "text", "list"))

Arguments

...
either a named list of parameters or a series of parameters to form the query

as
whether to return a data.frame, list, or text string nassqs_GET()

Value

a data frame, list, or text string of requested data.

See also

Examples

  # Get corn yields in Virginia in 2012
  params <- list(commodity_name = "CORN",
                 year = 2012,
                 agg_level_desc = "COUNTY",
                 state_alpha = "VA",
                 statisticcat_desc = "YIELD")
  yields <- nassqs(params)

#>   |                                                                              |                                                                      |   0%  |                                                                              |=                                                                     |   2%  |                                                                              |==                                                                    |   2%  |                                                                              |====                                                                  |   6%  |                                                                              |========                                                              |  11%  |                                                                              |===========                                                           |  16%  |                                                                              |============                                                          |  17%  |                                                                              |================                                                      |  22%  |                                                                              |===================                                                   |  27%  |                                                                              |====================                                                  |  28%  |                                                                              |=======================                                               |  33%  |                                                                              |===========================                                           |  39%  |                                                                              |============================                                          |  40%  |                                                                              |===============================                                       |  45%  |                                                                              |===================================                                   |  50%  |                                                                              |======================================                                |  55%  |                                                                              |==========================================                            |  60%  |                                                                              |==============================================                        |  65%  |                                                                              |=================================================                     |  70%  |                                                                              |=====================================================                 |  75%  |                                                                              |=====================================================                 |  76%  |                                                                              |=========================================================             |  81%  |                                                                              |=============================================================         |  87%  |                                                                              |================================================================      |  92%  |                                                                              |=================================================================     |  92%  |                                                                              |=================================================================     |  93%  |                                                                              |====================================================================  |  98%  |                                                                              |======================================================================| 100%

  head(yields)

#>   year congr_district_code        util_practice_desc asd_code state_ansi CV (%)
#> 1 2012                     ALL UTILIZATION PRACTICES       20         51       
#> 2 2012                     ALL UTILIZATION PRACTICES       20         51       
#> 3 2012                     ALL UTILIZATION PRACTICES       20         51       
#> 4 2012                     ALL UTILIZATION PRACTICES       20         51       
#> 5 2012                     ALL UTILIZATION PRACTICES       20         51       
#> 6 2012                     ALL UTILIZATION PRACTICES       20         51       
#>    class_desc agg_level_desc county_ansi                            short_desc
#> 1 ALL CLASSES         COUNTY         047 BARLEY - YIELD, MEASURED IN BU / ACRE
#> 2 ALL CLASSES         COUNTY         061 BARLEY - YIELD, MEASURED IN BU / ACRE
#> 3 ALL CLASSES         COUNTY         113 BARLEY - YIELD, MEASURED IN BU / ACRE
#> 4 ALL CLASSES         COUNTY         139 BARLEY - YIELD, MEASURED IN BU / ACRE
#> 5 ALL CLASSES         COUNTY         165 BARLEY - YIELD, MEASURED IN BU / ACRE
#> 6 ALL CLASSES         COUNTY         171 BARLEY - YIELD, MEASURED IN BU / ACRE
#>   Value region_desc domain_desc statisticcat_desc zip_5           load_time
#> 1  87.1                   TOTAL             YIELD       2015-06-02 15:00:27
#> 2  68.1                   TOTAL             YIELD       2015-06-02 15:00:27
#> 3  71.8                   TOTAL             YIELD       2015-06-02 15:00:27
#> 4    67                   TOTAL             YIELD       2015-06-02 15:00:27
#> 5    76                   TOTAL             YIELD       2015-06-02 15:00:27
#> 6    69                   TOTAL             YIELD       2015-06-02 15:00:27
#>   state_alpha  country_name                  location_desc sector_desc asd_desc
#> 1          VA UNITED STATES   VIRGINIA, NORTHERN, CULPEPER       CROPS NORTHERN
#> 2          VA UNITED STATES   VIRGINIA, NORTHERN, FAUQUIER       CROPS NORTHERN
#> 3          VA UNITED STATES    VIRGINIA, NORTHERN, MADISON       CROPS NORTHERN
#> 4          VA UNITED STATES       VIRGINIA, NORTHERN, PAGE       CROPS NORTHERN
#> 5          VA UNITED STATES VIRGINIA, NORTHERN, ROCKINGHAM       CROPS NORTHERN
#> 6          VA UNITED STATES VIRGINIA, NORTHERN, SHENANDOAH       CROPS NORTHERN
#>   state_fips_code begin_code watershed_desc county_code freq_desc
#> 1              51         00                        047    ANNUAL
#> 2              51         00                        061    ANNUAL
#> 3              51         00                        113    ANNUAL
#> 4              51         00                        139    ANNUAL
#> 5              51         00                        165    ANNUAL
#> 6              51         00                        171    ANNUAL
#>   reference_period_desc state_name week_ending      prodn_practice_desc
#> 1                  YEAR   VIRGINIA             ALL PRODUCTION PRACTICES
#> 2                  YEAR   VIRGINIA             ALL PRODUCTION PRACTICES
#> 3                  YEAR   VIRGINIA             ALL PRODUCTION PRACTICES
#> 4                  YEAR   VIRGINIA             ALL PRODUCTION PRACTICES
#> 5                  YEAR   VIRGINIA             ALL PRODUCTION PRACTICES
#> 6                  YEAR   VIRGINIA             ALL PRODUCTION PRACTICES
#>   source_desc unit_desc county_name watershed_code domaincat_desc country_code
#> 1      SURVEY BU / ACRE    CULPEPER       00000000  NOT SPECIFIED         9000
#> 2      SURVEY BU / ACRE    FAUQUIER       00000000  NOT SPECIFIED         9000
#> 3      SURVEY BU / ACRE     MADISON       00000000  NOT SPECIFIED         9000
#> 4      SURVEY BU / ACRE        PAGE       00000000  NOT SPECIFIED         9000
#> 5      SURVEY BU / ACRE  ROCKINGHAM       00000000  NOT SPECIFIED         9000
#> 6      SURVEY BU / ACRE  SHENANDOAH       00000000  NOT SPECIFIED         9000
#>    group_desc commodity_desc end_code
#> 1 FIELD CROPS         BARLEY       00
#> 2 FIELD CROPS         BARLEY       00
#> 3 FIELD CROPS         BARLEY       00
#> 4 FIELD CROPS         BARLEY       00
#> 5 FIELD CROPS         BARLEY       00
#> 6 FIELD CROPS         BARLEY       00