Get data and return a data frame

The primary function in the rnassqs package, nassqs makes a HTTP GET request to the USDA-NASS Quick Stats API and returns the data parsed as a data.frame, plain text, or list. Various other functions make use of nassqs to make specific queries. For a data request the Quick Stats API returns JSON that when parsed to a data.frame contains 39 columns and a varying number of rows depending on the query. Unfortunately there is not a way to restrict the number of columns.

nassqs(..., as = c("data.frame", "text", "list"))

Arguments

...
either a named list of parameters or a series of parameters to form the query

as
whether to return a data.frame, list, or text string nassqs_GET()

Value

a data frame, list, or text string of requested data.

See also

Examples

  # Get corn yields in Virginia in 2012
  params <- list(commodity_name = "CORN",
                 year = 2012,
                 agg_level_desc = "COUNTY",
                 state_alpha = "VA",
                 statisticcat_desc = "YIELD")
  yields <- nassqs(params)

#>   |                                                                              |                                                                      |   0%  |                                                                              |=                                                                     |   2%  |                                                                              |==                                                                    |   2%  |                                                                              |====                                                                  |   6%  |                                                                              |=======                                                               |  11%  |                                                                              |===========                                                           |  16%  |                                                                              |============                                                          |  17%  |                                                                              |===============                                                       |  22%  |                                                                              |===================                                                   |  27%  |                                                                              |====================                                                  |  28%  |                                                                              |=======================                                               |  33%  |                                                                              |===========================                                           |  38%  |                                                                              |===========================                                           |  39%  |                                                                              |===============================                                       |  44%  |                                                                              |===================================                                   |  49%  |                                                                              |===================================                                   |  50%  |                                                                              |=======================================                               |  55%  |                                                                              |==========================================                            |  60%  |                                                                              |==============================================                        |  66%  |                                                                              |==================================================                    |  71%  |                                                                              |==================================================                    |  72%  |                                                                              |======================================================                |  77%  |                                                                              |=========================================================             |  82%  |                                                                              |=============================================================         |  87%  |                                                                              |=============================================================         |  88%  |                                                                              |==============================================================        |  88%  |                                                                              |=================================================================     |  93%  |                                                                              |==================================================================    |  94%  |                                                                              |===================================================================== |  99%  |                                                                              |======================================================================| 100%

  head(yields)

#>                              short_desc county_name domain_desc region_desc
#> 1 BARLEY - YIELD, MEASURED IN BU / ACRE    CULPEPER       TOTAL            
#> 2 BARLEY - YIELD, MEASURED IN BU / ACRE    FAUQUIER       TOTAL            
#> 3 BARLEY - YIELD, MEASURED IN BU / ACRE     MADISON       TOTAL            
#> 4 BARLEY - YIELD, MEASURED IN BU / ACRE        PAGE       TOTAL            
#> 5 BARLEY - YIELD, MEASURED IN BU / ACRE  ROCKINGHAM       TOTAL            
#> 6 BARLEY - YIELD, MEASURED IN BU / ACRE  SHENANDOAH       TOTAL            
#>   domaincat_desc Value  group_desc zip_5 CV (%) county_ansi watershed_code
#> 1  NOT SPECIFIED  87.1 FIELD CROPS                      047       00000000
#> 2  NOT SPECIFIED  68.1 FIELD CROPS                      061       00000000
#> 3  NOT SPECIFIED  71.8 FIELD CROPS                      113       00000000
#> 4  NOT SPECIFIED    67 FIELD CROPS                      139       00000000
#> 5  NOT SPECIFIED    76 FIELD CROPS                      165       00000000
#> 6  NOT SPECIFIED    69 FIELD CROPS                      171       00000000
#>          util_practice_desc reference_period_desc source_desc asd_desc
#> 1 ALL UTILIZATION PRACTICES                  YEAR      SURVEY NORTHERN
#> 2 ALL UTILIZATION PRACTICES                  YEAR      SURVEY NORTHERN
#> 3 ALL UTILIZATION PRACTICES                  YEAR      SURVEY NORTHERN
#> 4 ALL UTILIZATION PRACTICES                  YEAR      SURVEY NORTHERN
#> 5 ALL UTILIZATION PRACTICES                  YEAR      SURVEY NORTHERN
#> 6 ALL UTILIZATION PRACTICES                  YEAR      SURVEY NORTHERN
#>   commodity_desc state_fips_code state_name sector_desc congr_district_code
#> 1         BARLEY              51   VIRGINIA       CROPS                    
#> 2         BARLEY              51   VIRGINIA       CROPS                    
#> 3         BARLEY              51   VIRGINIA       CROPS                    
#> 4         BARLEY              51   VIRGINIA       CROPS                    
#> 5         BARLEY              51   VIRGINIA       CROPS                    
#> 6         BARLEY              51   VIRGINIA       CROPS                    
#>   unit_desc state_ansi           load_time agg_level_desc  country_name
#> 1 BU / ACRE         51 2015-06-02 15:00:27         COUNTY UNITED STATES
#> 2 BU / ACRE         51 2015-06-02 15:00:27         COUNTY UNITED STATES
#> 3 BU / ACRE         51 2015-06-02 15:00:27         COUNTY UNITED STATES
#> 4 BU / ACRE         51 2015-06-02 15:00:27         COUNTY UNITED STATES
#> 5 BU / ACRE         51 2015-06-02 15:00:27         COUNTY UNITED STATES
#> 6 BU / ACRE         51 2015-06-02 15:00:27         COUNTY UNITED STATES
#>   asd_code      prodn_practice_desc state_alpha year watershed_desc
#> 1       20 ALL PRODUCTION PRACTICES          VA 2012               
#> 2       20 ALL PRODUCTION PRACTICES          VA 2012               
#> 3       20 ALL PRODUCTION PRACTICES          VA 2012               
#> 4       20 ALL PRODUCTION PRACTICES          VA 2012               
#> 5       20 ALL PRODUCTION PRACTICES          VA 2012               
#> 6       20 ALL PRODUCTION PRACTICES          VA 2012               
#>   country_code end_code  class_desc                  location_desc week_ending
#> 1         9000       00 ALL CLASSES   VIRGINIA, NORTHERN, CULPEPER            
#> 2         9000       00 ALL CLASSES   VIRGINIA, NORTHERN, FAUQUIER            
#> 3         9000       00 ALL CLASSES    VIRGINIA, NORTHERN, MADISON            
#> 4         9000       00 ALL CLASSES       VIRGINIA, NORTHERN, PAGE            
#> 5         9000       00 ALL CLASSES VIRGINIA, NORTHERN, ROCKINGHAM            
#> 6         9000       00 ALL CLASSES VIRGINIA, NORTHERN, SHENANDOAH            
#>   freq_desc statisticcat_desc county_code begin_code
#> 1    ANNUAL             YIELD         047         00
#> 2    ANNUAL             YIELD         061         00
#> 3    ANNUAL             YIELD         113         00
#> 4    ANNUAL             YIELD         139         00
#> 5    ANNUAL             YIELD         165         00
#> 6    ANNUAL             YIELD         171         00