Get data and return a data frame

The primary function in the rnassqs package, nassqs makes a HTTP GET request to the USDA-NASS Quick Stats API and returns the data parsed as a data.frame, plain text, or list. Various other functions make use of nassqs to make specific queries. For a data request the Quick Stats API returns JSON that when parsed to a data.frame contains 39 columns and a varying number of rows depending on the query. Unfortunately there is not a way to restrict the number of columns.

nassqs(..., as = c("data.frame", "text", "list"))

Arguments

...
either a named list of parameters or a series of parameters to form the query

as
whether to return a data.frame, list, or text string nassqs_GET()

Value

a data frame, list, or text string of requested data.

See also

Examples

  # Get corn yields in Virginia in 2012
  params <- list(commodity_name = "CORN",
                 year = 2012,
                 agg_level_desc = "COUNTY",
                 state_alpha = "VA",
                 statisticcat_desc = "YIELD")
  yields <- nassqs(params)

#>   |                                                                              |                                                                      |   0%  |                                                                              |=                                                                     |   2%  |                                                                              |==                                                                    |   2%  |                                                                              |===                                                                   |   5%  |                                                                              |=====                                                                 |   7%  |                                                                              |=========                                                             |  13%  |                                                                              |============                                                          |  18%  |                                                                              |=============                                                         |  18%  |                                                                              |=============                                                         |  19%  |                                                                              |=================                                                     |  24%  |                                                                              |=================                                                     |  25%  |                                                                              |=====================                                                 |  30%  |                                                                              |=====================                                                 |  31%  |                                                                              |=========================                                             |  36%  |                                                                              |==========================                                            |  37%  |                                                                              |=============================                                         |  42%  |                                                                              |=================================                                     |  47%  |                                                                              |==================================                                    |  48%  |                                                                              |=====================================                                 |  53%  |                                                                              |=====================================                                 |  54%  |                                                                              |======================================                                |  54%  |                                                                              |=========================================                             |  59%  |                                                                              |=============================================                         |  64%  |                                                                              |=============================================                         |  65%  |                                                                              |==============================================                        |  65%  |                                                                              |=================================================                     |  70%  |                                                                              |==================================================                    |  71%  |                                                                              |=====================================================                 |  76%  |                                                                              |======================================================                |  77%  |                                                                              |==========================================================            |  82%  |                                                                              |==========================================================            |  83%  |                                                                              |==============================================================        |  88%  |                                                                              |==================================================================    |  94%  |                                                                              |===================================================================== |  99%  |                                                                              |======================================================================| 100%

  head(yields)

#>   state_name week_ending      prodn_practice_desc source_desc unit_desc
#> 1   VIRGINIA             ALL PRODUCTION PRACTICES      SURVEY BU / ACRE
#> 2   VIRGINIA             ALL PRODUCTION PRACTICES      SURVEY BU / ACRE
#> 3   VIRGINIA             ALL PRODUCTION PRACTICES      SURVEY BU / ACRE
#> 4   VIRGINIA             ALL PRODUCTION PRACTICES      SURVEY BU / ACRE
#> 5   VIRGINIA             ALL PRODUCTION PRACTICES      SURVEY BU / ACRE
#> 6   VIRGINIA             ALL PRODUCTION PRACTICES      SURVEY BU / ACRE
#>   state_fips_code begin_code watershed_desc county_code freq_desc
#> 1              51         00                        047    ANNUAL
#> 2              51         00                        061    ANNUAL
#> 3              51         00                        113    ANNUAL
#> 4              51         00                        139    ANNUAL
#> 5              51         00                        165    ANNUAL
#> 6              51         00                        171    ANNUAL
#>   reference_period_desc  group_desc commodity_desc end_code county_name
#> 1                  YEAR FIELD CROPS         BARLEY       00    CULPEPER
#> 2                  YEAR FIELD CROPS         BARLEY       00    FAUQUIER
#> 3                  YEAR FIELD CROPS         BARLEY       00     MADISON
#> 4                  YEAR FIELD CROPS         BARLEY       00        PAGE
#> 5                  YEAR FIELD CROPS         BARLEY       00  ROCKINGHAM
#> 6                  YEAR FIELD CROPS         BARLEY       00  SHENANDOAH
#>   watershed_code domaincat_desc country_code asd_code state_ansi CV (%)
#> 1       00000000  NOT SPECIFIED         9000       20         51       
#> 2       00000000  NOT SPECIFIED         9000       20         51       
#> 3       00000000  NOT SPECIFIED         9000       20         51       
#> 4       00000000  NOT SPECIFIED         9000       20         51       
#> 5       00000000  NOT SPECIFIED         9000       20         51       
#> 6       00000000  NOT SPECIFIED         9000       20         51       
#>   agg_level_desc  class_desc county_ansi                            short_desc
#> 1         COUNTY ALL CLASSES         047 BARLEY - YIELD, MEASURED IN BU / ACRE
#> 2         COUNTY ALL CLASSES         061 BARLEY - YIELD, MEASURED IN BU / ACRE
#> 3         COUNTY ALL CLASSES         113 BARLEY - YIELD, MEASURED IN BU / ACRE
#> 4         COUNTY ALL CLASSES         139 BARLEY - YIELD, MEASURED IN BU / ACRE
#> 5         COUNTY ALL CLASSES         165 BARLEY - YIELD, MEASURED IN BU / ACRE
#> 6         COUNTY ALL CLASSES         171 BARLEY - YIELD, MEASURED IN BU / ACRE
#>   Value year congr_district_code        util_practice_desc  country_name
#> 1  87.1 2012                     ALL UTILIZATION PRACTICES UNITED STATES
#> 2  68.1 2012                     ALL UTILIZATION PRACTICES UNITED STATES
#> 3  71.8 2012                     ALL UTILIZATION PRACTICES UNITED STATES
#> 4    67 2012                     ALL UTILIZATION PRACTICES UNITED STATES
#> 5    76 2012                     ALL UTILIZATION PRACTICES UNITED STATES
#> 6    69 2012                     ALL UTILIZATION PRACTICES UNITED STATES
#>                    location_desc sector_desc asd_desc region_desc domain_desc
#> 1   VIRGINIA, NORTHERN, CULPEPER       CROPS NORTHERN                   TOTAL
#> 2   VIRGINIA, NORTHERN, FAUQUIER       CROPS NORTHERN                   TOTAL
#> 3    VIRGINIA, NORTHERN, MADISON       CROPS NORTHERN                   TOTAL
#> 4       VIRGINIA, NORTHERN, PAGE       CROPS NORTHERN                   TOTAL
#> 5 VIRGINIA, NORTHERN, ROCKINGHAM       CROPS NORTHERN                   TOTAL
#> 6 VIRGINIA, NORTHERN, SHENANDOAH       CROPS NORTHERN                   TOTAL
#>   zip_5 statisticcat_desc           load_time state_alpha
#> 1                   YIELD 2015-06-02 15:00:27          VA
#> 2                   YIELD 2015-06-02 15:00:27          VA
#> 3                   YIELD 2015-06-02 15:00:27          VA
#> 4                   YIELD 2015-06-02 15:00:27          VA
#> 5                   YIELD 2015-06-02 15:00:27          VA
#> 6                   YIELD 2015-06-02 15:00:27          VA