Get data and return a data frame

The primary function in the rnassqs package, nassqs makes a HTTP GET request to the USDA-NASS Quick Stats API and returns the data parsed as a data.frame, plain text, or list. Various other functions make use of nassqs to make specific queries. For a data request the Quick Stats API returns JSON that when parsed to a data.frame contains 39 columns and a varying number of rows depending on the query. Unfortunately there is not a way to restrict the number of columns.

nassqs(..., as = c("data.frame", "text", "list"))

Arguments

...
either a named list of parameters or a series of parameters to form the query

as
whether to return a data.frame, list, or text string nassqs_GET()

Value

a data frame, list, or text string of requested data.

See also

Examples

  # Get corn yields in Virginia in 2012
  params <- list(commodity_name = "CORN",
                 year = 2012,
                 agg_level_desc = "COUNTY",
                 state_alpha = "VA",
                 statisticcat_desc = "YIELD")
  yields <- nassqs(params)

#>   |                                                                              |                                                                      |   0%  |                                                                              |=                                                                     |   2%  |                                                                              |==                                                                    |   2%  |                                                                              |===                                                                   |   5%  |                                                                              |=====                                                                 |   7%  |                                                                              |=========                                                             |  13%  |                                                                              |============                                                          |  18%  |                                                                              |=============                                                         |  18%  |                                                                              |=============                                                         |  19%  |                                                                              |=================                                                     |  24%  |                                                                              |====================                                                  |  29%  |                                                                              |=====================                                                 |  30%  |                                                                              |========================                                              |  35%  |                                                                              |============================                                          |  40%  |                                                                              |=============================                                         |  41%  |                                                                              |================================                                      |  46%  |                                                                              |====================================                                  |  51%  |                                                                              |====================================                                  |  52%  |                                                                              |========================================                              |  57%  |                                                                              |============================================                          |  62%  |                                                                              |===============================================                       |  67%  |                                                                              |===================================================                   |  73%  |                                                                              |======================================================                |  78%  |                                                                              |==========================================================            |  83%  |                                                                              |==============================================================        |  88%  |                                                                              |=================================================================     |  93%  |                                                                              |==================================================================    |  94%  |                                                                              |===================================================================== |  99%  |                                                                              |======================================================================| 100%

  head(yields)

#>   source_desc domain_desc state_alpha reference_period_desc  country_name
#> 1      SURVEY       TOTAL          VA                  YEAR UNITED STATES
#> 2      SURVEY       TOTAL          VA                  YEAR UNITED STATES
#> 3      SURVEY       TOTAL          VA                  YEAR UNITED STATES
#> 4      SURVEY       TOTAL          VA                  YEAR UNITED STATES
#> 5      SURVEY       TOTAL          VA                  YEAR UNITED STATES
#> 6      SURVEY       TOTAL          VA                  YEAR UNITED STATES
#>   freq_desc CV (%) end_code                            short_desc
#> 1    ANNUAL              00 BARLEY - YIELD, MEASURED IN BU / ACRE
#> 2    ANNUAL              00 BARLEY - YIELD, MEASURED IN BU / ACRE
#> 3    ANNUAL              00 BARLEY - YIELD, MEASURED IN BU / ACRE
#> 4    ANNUAL              00 BARLEY - YIELD, MEASURED IN BU / ACRE
#> 5    ANNUAL              00 BARLEY - YIELD, MEASURED IN BU / ACRE
#> 6    ANNUAL              00 BARLEY - YIELD, MEASURED IN BU / ACRE
#>   congr_district_code week_ending      prodn_practice_desc  class_desc
#> 1                                 ALL PRODUCTION PRACTICES ALL CLASSES
#> 2                                 ALL PRODUCTION PRACTICES ALL CLASSES
#> 3                                 ALL PRODUCTION PRACTICES ALL CLASSES
#> 4                                 ALL PRODUCTION PRACTICES ALL CLASSES
#> 5                                 ALL PRODUCTION PRACTICES ALL CLASSES
#> 6                                 ALL PRODUCTION PRACTICES ALL CLASSES
#>   unit_desc agg_level_desc watershed_code country_code sector_desc  group_desc
#> 1 BU / ACRE         COUNTY       00000000         9000       CROPS FIELD CROPS
#> 2 BU / ACRE         COUNTY       00000000         9000       CROPS FIELD CROPS
#> 3 BU / ACRE         COUNTY       00000000         9000       CROPS FIELD CROPS
#> 4 BU / ACRE         COUNTY       00000000         9000       CROPS FIELD CROPS
#> 5 BU / ACRE         COUNTY       00000000         9000       CROPS FIELD CROPS
#> 6 BU / ACRE         COUNTY       00000000         9000       CROPS FIELD CROPS
#>   year region_desc watershed_desc begin_code Value commodity_desc asd_code
#> 1 2012                                    00  87.1         BARLEY       20
#> 2 2012                                    00  68.1         BARLEY       20
#> 3 2012                                    00  71.8         BARLEY       20
#> 4 2012                                    00    67         BARLEY       20
#> 5 2012                                    00    76         BARLEY       20
#> 6 2012                                    00    69         BARLEY       20
#>   county_ansi        util_practice_desc asd_desc county_name
#> 1         047 ALL UTILIZATION PRACTICES NORTHERN    CULPEPER
#> 2         061 ALL UTILIZATION PRACTICES NORTHERN    FAUQUIER
#> 3         113 ALL UTILIZATION PRACTICES NORTHERN     MADISON
#> 4         139 ALL UTILIZATION PRACTICES NORTHERN        PAGE
#> 5         165 ALL UTILIZATION PRACTICES NORTHERN  ROCKINGHAM
#> 6         171 ALL UTILIZATION PRACTICES NORTHERN  SHENANDOAH
#>                    location_desc           load_time county_code
#> 1   VIRGINIA, NORTHERN, CULPEPER 2015-06-02 15:00:27         047
#> 2   VIRGINIA, NORTHERN, FAUQUIER 2015-06-02 15:00:27         061
#> 3    VIRGINIA, NORTHERN, MADISON 2015-06-02 15:00:27         113
#> 4       VIRGINIA, NORTHERN, PAGE 2015-06-02 15:00:27         139
#> 5 VIRGINIA, NORTHERN, ROCKINGHAM 2015-06-02 15:00:27         165
#> 6 VIRGINIA, NORTHERN, SHENANDOAH 2015-06-02 15:00:27         171
#>   statisticcat_desc state_ansi domaincat_desc zip_5 state_fips_code state_name
#> 1             YIELD         51  NOT SPECIFIED                    51   VIRGINIA
#> 2             YIELD         51  NOT SPECIFIED                    51   VIRGINIA
#> 3             YIELD         51  NOT SPECIFIED                    51   VIRGINIA
#> 4             YIELD         51  NOT SPECIFIED                    51   VIRGINIA
#> 5             YIELD         51  NOT SPECIFIED                    51   VIRGINIA
#> 6             YIELD         51  NOT SPECIFIED                    51   VIRGINIA