Issue a GET request to the NASS 'Quick Stats' API

This is the workhorse of the package that provides the core request functionality to the NASS 'Quick Stats' API: https://quickstats.nass.usda.gov/api. In most cases nassqs() or other high-level functions should be used. nassqs_GET() uses httr::GET() to make a HTTP GET request, which returns a request object which must then be parsed to a data.frame, list, or other R object. Higher-level functions will do that parsing automatically. However, if you need access to the request object directly, nassqs_GET() provides that.

Usage

nassqs_GET(
  ...,
  api_path = c("api_GET", "get_param_values", "get_counts"),
  progress_bar = TRUE,
  format = c("csv", "json", "xml")
)

Arguments

...

either a named list of parameters or a series of parameters to use in the query

api_path

the API path that determines the type of request being made.

progress_bar

whether to display the project bar or not.

format

The format to return the query in. Only useful if as = "text".

Value

a httr::GET() response object

Examples

# \donttest{
  # Yields for corn in 2012 in Washington
  params <- list(commodity_desc = "CORN",
                 year = 2012,
                 agg_level_desc = "STATE",
                 state_alpha = "WA",
                 statisticcat_desc = "YIELD")

  # Returns a request object that must be parsed either manually or
  # by using nassqs_parse()
  response <- nassqs_GET(params)
#>   |                                                                              |                                                                      |   0%  |                                                                              |===================================================================   |  96%  |                                                                              |======================================================================| 100%
  yields <- nassqs_parse(response)
  head(yields)
#>   source_desc sector_desc  group_desc commodity_desc  class_desc
#> 1      CENSUS       CROPS FIELD CROPS           CORN ALL CLASSES
#> 2      CENSUS       CROPS FIELD CROPS           CORN ALL CLASSES
#> 3      CENSUS       CROPS FIELD CROPS           CORN ALL CLASSES
#> 4      CENSUS       CROPS FIELD CROPS           CORN ALL CLASSES
#> 5      CENSUS       CROPS FIELD CROPS           CORN ALL CLASSES
#> 6      CENSUS       CROPS FIELD CROPS           CORN ALL CLASSES
#>       prodn_practice_desc util_practice_desc statisticcat_desc   unit_desc
#> 1  IRRIGATED, ENTIRE CROP              GRAIN             YIELD   BU / ACRE
#> 2  IRRIGATED, ENTIRE CROP             SILAGE             YIELD TONS / ACRE
#> 3 IRRIGATED, NONE OF CROP              GRAIN             YIELD   BU / ACRE
#> 4 IRRIGATED, NONE OF CROP             SILAGE             YIELD TONS / ACRE
#> 5 IRRIGATED, PART OF CROP              GRAIN             YIELD   BU / ACRE
#> 6 IRRIGATED, PART OF CROP             SILAGE             YIELD TONS / ACRE
#>                                                               short_desc
#> 1     CORN, GRAIN, IRRIGATED, ENTIRE CROP - YIELD, MEASURED IN BU / ACRE
#> 2  CORN, SILAGE, IRRIGATED, ENTIRE CROP - YIELD, MEASURED IN TONS / ACRE
#> 3    CORN, GRAIN, IRRIGATED, NONE OF CROP - YIELD, MEASURED IN BU / ACRE
#> 4 CORN, SILAGE, IRRIGATED, NONE OF CROP - YIELD, MEASURED IN TONS / ACRE
#> 5    CORN, GRAIN, IRRIGATED, PART OF CROP - YIELD, MEASURED IN BU / ACRE
#> 6 CORN, SILAGE, IRRIGATED, PART OF CROP - YIELD, MEASURED IN TONS / ACRE
#>   domain_desc domaincat_desc agg_level_desc state_ansi state_fips_code
#> 1       TOTAL  NOT SPECIFIED          STATE         53              53
#> 2       TOTAL  NOT SPECIFIED          STATE         53              53
#> 3       TOTAL  NOT SPECIFIED          STATE         53              53
#> 4       TOTAL  NOT SPECIFIED          STATE         53              53
#> 5       TOTAL  NOT SPECIFIED          STATE         53              53
#> 6       TOTAL  NOT SPECIFIED          STATE         53              53
#>   state_alpha state_name asd_code asd_desc county_ansi county_code county_name
#> 1          WA WASHINGTON                                                      
#> 2          WA WASHINGTON                                                      
#> 3          WA WASHINGTON                                                      
#> 4          WA WASHINGTON                                                      
#> 5          WA WASHINGTON                                                      
#> 6          WA WASHINGTON                                                      
#>   region_desc zip_5 watershed_code watershed_desc congr_district_code
#> 1                         00000000                                   
#> 2                         00000000                                   
#> 3                         00000000                                   
#> 4                         00000000                                   
#> 5                         00000000                                   
#> 6                         00000000                                   
#>   country_code  country_name location_desc year freq_desc begin_code end_code
#> 1         9000 UNITED STATES    WASHINGTON 2012    ANNUAL         00       00
#> 2         9000 UNITED STATES    WASHINGTON 2012    ANNUAL         00       00
#> 3         9000 UNITED STATES    WASHINGTON 2012    ANNUAL         00       00
#> 4         9000 UNITED STATES    WASHINGTON 2012    ANNUAL         00       00
#> 5         9000 UNITED STATES    WASHINGTON 2012    ANNUAL         00       00
#> 6         9000 UNITED STATES    WASHINGTON 2012    ANNUAL         00       00
#>   reference_period_desc week_ending           load_time Value CV (%)
#> 1                  YEAR             2012-12-31 00:00:00    NA    (D)
#> 2                  YEAR             2012-12-31 00:00:00  26.3    0.9
#> 3                  YEAR             2012-12-31 00:00:00    NA    (D)
#> 4                  YEAR             2012-12-31 00:00:00  21.9    0.7
#> 5                  YEAR             2012-12-31 00:00:00    NA    (D)
#> 6                  YEAR             2012-12-31 00:00:00  20.7    0.7

  # Get the number of records that would be returned for a given request
  # Equivalent to 'nassqs_record_count(params)'
  response <- nassqs_GET(params, api_path = "get_counts")
#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%
  records <- nassqs_parse(response)
  records
#> $count
#> [1] 12
#> 
#> $Value
#> numeric(0)
#> 

  # Get the list of allowable values for the parameters 'statisticcat_desc'
  # Equivalent to 'nassqs_param_values("statisticcat_desc")'
  req <- nassqs_GET(list(param = "statisticcat_desc"),
                    api_path = "get_param_values")
#>   |                                                                              |                                                                      |   0%  |                                                                              |=================================================================     |  93%  |                                                                              |======================================================================| 100%
  statisticcat_desc_values <- nassqs_parse(req, as = "list")
  head(statisticcat_desc_values)
#> $statisticcat_desc
#>   [1] "ACCESSIBILITY"                                    
#>   [2] "ACCESSIBILITY, 5 YEAR AVG"                        
#>   [3] "ACCESSIBILITY, PREVIOUS YEAR"                     
#>   [4] "ACTIVE GINS"                                      
#>   [5] "ACTIVITY"                                         
#>   [6] "ACTIVITY, 5 YEAR AVG"                             
#>   [7] "ACTIVITY, PREVIOUS YEAR"                          
#>   [8] "ADDED & REPLACED"                                 
#>   [9] "AGE, AVG"                                         
#>  [10] "APPLICATIONS"                                     
#>  [11] "AREA"                                             
#>  [12] "AREA BEARING"                                     
#>  [13] "AREA BEARING & NON-BEARING"                       
#>  [14] "AREA BEARING & NONBEARING"                        
#>  [15] "AREA CERTIFIED"                                   
#>  [16] "AREA CROPLAND"                                    
#>  [17] "AREA FILLED"                                      
#>  [18] "AREA FLOODED"                                     
#>  [19] "AREA GRAZED"                                      
#>  [20] "AREA GROWN"                                       
#>  [21] "AREA HARVESTED"                                   
#>  [22] "AREA IN PRODUCTION"                               
#>  [23] "AREA IN PRODUCTION PREVIOUS CENSUS YEAR"          
#>  [24] "AREA IRRIGATED"                                   
#>  [25] "AREA IRRIGATED PREVIOUS CENSUS YEAR"              
#>  [26] "AREA NON-BEARING"                                 
#>  [27] "AREA NOT HARVESTED"                               
#>  [28] "AREA OF WATER"                                    
#>  [29] "AREA OPERATED"                                    
#>  [30] "AREA PLANTED"                                     
#>  [31] "AREA PLANTED, NET"                                
#>  [32] "AREA POLLINATED, PAID BASIS"                      
#>  [33] "AREA REMAINING TO BE PLANTED"                     
#>  [34] "AREA RENTED"                                      
#>  [35] "ASSET VALUE"                                      
#>  [36] "AVAILABILITY"                                     
#>  [37] "AVAILABILITY, 5 YEAR AVG"                         
#>  [38] "AVAILABILITY, PREVIOUS YEAR"                      
#>  [39] "BEING MOLTED"                                     
#>  [40] "BLOOM DATE"                                       
#>  [41] "BOLL COUNT"                                       
#>  [42] "BRED"                                             
#>  [43] "BROKEN"                                           
#>  [44] "CAGES"                                            
#>  [45] "CALF CROP"                                        
#>  [46] "CAPACITY"                                         
#>  [47] "CAPTURED"                                         
#>  [48] "CLIPPED"                                          
#>  [49] "CONDEMNED, ANTE-MORTEM"                           
#>  [50] "CONDEMNED, POST-MORTEM"                           
#>  [51] "CONDEMNED, POST-MORTEM, AIRSACCULITIS"            
#>  [52] "CONDEMNED, POST-MORTEM, ALL CAUSES"               
#>  [53] "CONDEMNED, POST-MORTEM, BRUISES"                  
#>  [54] "CONDEMNED, POST-MORTEM, CADAVER"                  
#>  [55] "CONDEMNED, POST-MORTEM, CONTAMINATION"            
#>  [56] "CONDEMNED, POST-MORTEM, LEUKOSIS"                 
#>  [57] "CONDEMNED, POST-MORTEM, OTHER REASONS"            
#>  [58] "CONDEMNED, POST-MORTEM, OVERSCALD"                
#>  [59] "CONDEMNED, POST-MORTEM, SEPTICAEMIA"              
#>  [60] "CONDEMNED, POST-MORTEM, SYNOVITIS"                
#>  [61] "CONDEMNED, POST-MORTEM, TUBERCULOSIS"             
#>  [62] "CONDEMNED, POST-MORTEM, TUMORS"                   
#>  [63] "CONDEMNED, POST-MORTEM, UNKNOWN CAUSES"           
#>  [64] "CONDITION"                                        
#>  [65] "CONDITION, 5 YEAR AVG"                            
#>  [66] "CONDITION, PREVIOUS YEAR"                         
#>  [67] "CONDUCT BUSINESS"                                 
#>  [68] "COVER"                                            
#>  [69] "COVER, PREVIOUS YEAR"                             
#>  [70] "CRUSHED"                                          
#>  [71] "DAMAGE"                                           
#>  [72] "DAMAGE, 5 YEAR AVG"                               
#>  [73] "DAMAGE, PREVIOUS YEAR"                            
#>  [74] "DAYS SUITABLE"                                    
#>  [75] "DAYS SUITABLE, PREVIOUS YEAR"                     
#>  [76] "DEBT"                                             
#>  [77] "DELIVERED VALUE"                                  
#>  [78] "DEPTH TO BOWLS"                                   
#>  [79] "DEPTH TO WATER"                                   
#>  [80] "DEPTH, AVG"                                       
#>  [81] "DEPTH, AVG, PREVIOUS YEAR"                        
#>  [82] "DEVICES"                                          
#>  [83] "DISAPPEARANCE"                                    
#>  [84] "DISAPPEARANCE, OTHER"                             
#>  [85] "DISTRIBUTION"                                     
#>  [86] "EAR COUNT"                                        
#>  [87] "EGGS IN INCUBATORS"                               
#>  [88] "EGGS SET"                                         
#>  [89] "END DATE, AVG"                                    
#>  [90] "END DATE, MAX"                                    
#>  [91] "ENGINE SIZE"                                      
#>  [92] "EXPENSE"                                          
#>  [93] "FACILITIES"                                       
#>  [94] "FAMILIES"                                         
#>  [95] "FARM OPERATIONS SERVICED"                         
#>  [96] "FARM USE"                                         
#>  [97] "FARROWED"                                         
#>  [98] "FAT TEST"                                         
#>  [99] "FEEDLOTS"                                         
#> [100] "FRUIT SIZE"                                       
#> [101] "FRUIT SIZE, 5 YEAR AVG"                           
#> [102] "FRUIT SIZE, PREVIOUS YEAR"                        
#> [103] "GAIN"                                             
#> [104] "GINNED BALES"                                     
#> [105] "GROSS INCOME"                                     
#> [106] "HARVEST LOSS"                                     
#> [107] "HATCHED"                                          
#> [108] "HEAD COUNT"                                       
#> [109] "HEIGHT, AVG"                                      
#> [110] "HEIGHT, AVG, 5 YEAR AVG"                          
#> [111] "HEIGHT, AVG, PREVIOUS YEAR"                       
#> [112] "HOURS OF OPERATION"                               
#> [113] "HOUSEHOLD USE"                                    
#> [114] "INDEX FOR PRICE PAID, 1910 - 1914"                
#> [115] "INDEX FOR PRICE PAID, 1964 - 1968 OVERALL DERIVED"
#> [116] "INDEX FOR PRICE PAID, 1990 - 1992"                
#> [117] "INDEX FOR PRICE PAID, 2011"                       
#> [118] "INDEX FOR PRICE RECEIVED, 1910 - 1914"            
#> [119] "INDEX FOR PRICE RECEIVED, 1990 - 1992"            
#> [120] "INDEX FOR PRICE RECEIVED, 2011"                   
#> [121] "INFESTATION"                                      
#> [122] "INFESTATION, PREVIOUS YEAR"                       
#> [123] "INVENTORY"                                        
#> [124] "INVENTORY, AVG"                                   
#> [125] "INVENTORY, MAX"                                   
#> [126] "INVENTORY, PEAK"                                  
#> [127] "KID CROP"                                         
#> [128] "LAMB CROP"                                        
#> [129] "LAMBING RATE"                                     
#> [130] "LANDLORDS"                                        
#> [131] "LENGTH OF SEASON, AVG"                            
#> [132] "LITTER RATE"                                      
#> [133] "LOSS"                                             
#> [134] "LOSS, COLONY COLLAPSE DISORDER"                   
#> [135] "LOSS, DEADOUT"                                    
#> [136] "LOSS, DEATH"                                      
#> [137] "LOSS, DEATH & RENDERED"                           
#> [138] "LOSS, DEATH, 5 YEAR AVG"                          
#> [139] "LOSS, DEATH, PREVIOUS YEAR"                       
#> [140] "MILKFAT"                                          
#> [141] "MILL SALE"                                        
#> [142] "MILLED"                                           
#> [143] "MILLING CAPACITY"                                 
#> [144] "MOISTURE"                                         
#> [145] "MOISTURE CONTENT"                                 
#> [146] "MOISTURE, 5 YEAR AVG"                             
#> [147] "MOISTURE, PREVIOUS YEAR"                          
#> [148] "MOLT COMPLETED"                                   
#> [149] "MOVEMENT"                                         
#> [150] "MOVEMENT, 5 YEAR AVG"                             
#> [151] "MOVEMENT, PREVIOUS YEAR"                          
#> [152] "NET INCOME"                                       
#> [153] "NET WEIGHT"                                       
#> [154] "NUT SET"                                          
#> [155] "NUT SET, 5 YEAR AVG"                              
#> [156] "NUT SET, PREVIOUS YEAR"                           
#> [157] "OCEAN HARVESTED"                                  
#> [158] "OPERATING PRESSURE"                               
#> [159] "OPERATIONS"                                       
#> [160] "OPERATORS"                                        
#> [161] "ORGANIZATIONS"                                    
#> [162] "OTHER SALE"                                       
#> [163] "PARTNERS, AVG"                                    
#> [164] "PENS"                                             
#> [165] "PERSONS IN HOUSEHOLD"                             
#> [166] "PEST MGMT"                                        
#> [167] "PIG CROP"                                         
#> [168] "PLACEMENTS"                                       
#> [169] "PLACEMENTS, INTENDED"                             
#> [170] "PLACEMENTS, NET"                                  
#> [171] "PLANT POPULATION"                                 
#> [172] "POD COUNT"                                        
#> [173] "POLLINATION"                                      
#> [174] "PONDS"                                            
#> [175] "PONDS & TANKS"                                    
#> [176] "PRICE PAID"                                       
#> [177] "PRICE REACTION"                                   
#> [178] "PRICE REACTION, DECREASE"                         
#> [179] "PRICE REACTION, INCREASE"                         
#> [180] "PRICE REACTION, NO CHANGE"                        
#> [181] "PRICE RECEIVED"                                   
#> [182] "PRICE RECEIVED AFTER REPORT"                      
#> [183] "PRICE RECEIVED PRIOR TO CLOSING"                  
#> [184] "PRICE RECEIVED, 3 YEAR AVG"                       
#> [185] "PRICE RECEIVED, ADJUSTED BASE"                    
#> [186] "PRICE RECEIVED, PARITY"                           
#> [187] "PRICE RECEIVED, PREMIUM"                          
#> [188] "PRICE RECEIVED, SEASONALLY ADJUSTED"              
#> [189] "PROCESSED IN LAB"                                 
#> [190] "PRODUCERS"                                        
#> [191] "PRODUCTION"                                       
#> [192] "PRODUCTION NOT SOLD"                              
#> [193] "PROGRESS"                                         
#> [194] "PROGRESS, 5 YEAR AVG"                             
#> [195] "PROGRESS, PREVIOUS YEAR"                          
#> [196] "PUMPING CAPACITY"                                 
#> [197] "PUMPING LIFT"                                     
#> [198] "PUMPS"                                            
#> [199] "RACEWAYS"                                         
#> [200] "RATE OF LAY"                                      
#> [201] "RATIO"                                            
#> [202] "RECEIPTS"                                         
#> [203] "RECEIVING"                                        
#> [204] "RECEIVING, 5 YEAR AVG"                            
#> [205] "RECEIVING, PREVIOUS YEAR"                         
#> [206] "RELATIVE WEIGHT"                                  
#> [207] "RELATIVE WEIGHT, 5 YEAR MOVING AVERAGE"           
#> [208] "REMOVAL FOR PROCESSING"                           
#> [209] "REMOVAL FOR PROCESSING, EDIBLE USE"               
#> [210] "REMOVAL FOR PROCESSING, INEDIBLE USE"             
#> [211] "REPAYMENTS"                                       
#> [212] "ROW WIDTH"                                        
#> [213] "SALES"                                            
#> [214] "SALES & DISTRIBUTION"                             
#> [215] "SALES FOR SLAUGHTER"                              
#> [216] "SALES IN CONVENTIONAL MARKETS"                    
#> [217] "SALES IN ORGANIC MARKETS"                         
#> [218] "SAMPLES"                                          
#> [219] "SAVINGS"                                          
#> [220] "SEASONAL ADJUSTMENT FACTOR FOR PRICE RECEIVED"    
#> [221] "SEED FOR PLANTING"                                
#> [222] "SHIPMENTS"                                        
#> [223] "SHIPMENTS IN"                                     
#> [224] "SHORN"                                            
#> [225] "SHRINK"                                           
#> [226] "SLAUGHTERED"                                      
#> [227] "SOURCE"                                           
#> [228] "SOURCE, 5 YEAR AVG"                               
#> [229] "SOURCE, PREVIOUS YEAR"                            
#> [230] "SPINDLES"                                         
#> [231] "START DATE"                                       
#> [232] "START DATE, 5 YEAR AVG"                           
#> [233] "START DATE, AVG"                                  
#> [234] "START DATE, MIN"                                  
#> [235] "START DATE, PREVIOUS YEAR"                        
#> [236] "STOCKS"                                           
#> [237] "STOCKS, CURRENT YEAR"                             
#> [238] "STOCKS, PREVIOUS YEAR"                            
#> [239] "SUCROSE"                                          
#> [240] "SUPPLIES"                                         
#> [241] "SUPPLIES, 5 YEAR AVG"                             
#> [242] "SUPPLIES, PREVIOUS YEAR"                          
#> [243] "TANKS"                                            
#> [244] "TAPS"                                             
#> [245] "TENANTS"                                          
#> [246] "TIME OPERATED"                                    
#> [247] "TIME WORKED"                                      
#> [248] "TIME WORKED, BASE"                                
#> [249] "TOTAL AREA"                                       
#> [250] "TREATED"                                          
#> [251] "USAGE"                                            
#> [252] "VALUE"                                            
#> [253] "VALUE OF POLLINATION"                             
#> [254] "VOLUME"                                           
#> [255] "VOLUME HANDLED"                                   
#> [256] "WAGE RATE"                                        
#> [257] "WAGE RATE, BASE"                                  
#> [258] "WAREHOUSES"                                       
#> [259] "WATER APPLIED"                                    
#> [260] "WATER DELIVERED"                                  
#> [261] "WATER RECEIVED"                                   
#> [262] "WELL DEPTH"                                       
#> [263] "WELLS"                                            
#> [264] "WORKERS"                                          
#> [265] "WORKERS, AVG"                                     
#> [266] "WORKERS, PEAK"                                    
#> [267] "YEARS ON ANY OPERATION, AVG"                      
#> [268] "YEARS ON PRESENT OPERATION, AVG"                  
#> [269] "YEARS RENTED TO TENANT, AVG"                      
#> [270] "YIELD"                                            
#> 
#> $Value
#> numeric(0)
#> 
# }