Issue a GET request to the NASS 'Quick Stats' API

This is the workhorse of the package that provides the core request functionality to the NASS 'Quick Stats' API: https://quickstats.nass.usda.gov/api. In most cases nassqs() or other high-level functions should be used. nassqs_GET uses httr::GET() to make a HTTP GET request, which returns a request object which must then be parsed to a data.frame, list, or other R object. Higher-level functions will do that parsing automatically. However, if you need access to the request object directly, nassqs_GET provides that.

nassqs_GET(..., api_path = c("api_GET", "get_param_values",
  "get_counts"))

Arguments

...
either a named list of parameters or a series of parameters to use in the query

api_path
the API path that determines the type of request being made

Value

a httr::GET() response object

Examples

  # Yields for corn in 2012 in Washington
  params <- list(commodity_name = "CORN",
                 year = 2012,
                 agg_level_desc = "STATE",
                 state_alpha = "WA",
                 statisticcat_desc = "YIELD")

  # Returns a request object that must be parsed either manually or
  # by using nassqs_parse()
  response <- nassqs_GET(params)

#>   |                                                                              |                                                                      |   0%  |                                                                              |=====                                                                 |   7%  |                                                                              |======                                                                |   8%  |                                                                              |===========                                                           |  15%  |                                                                              |========================                                              |  34%  |                                                                              |=====================================                                 |  52%  |                                                                              |==================================================                    |  71%  |                                                                              |===============================================================       |  90%  |                                                                              |================================================================      |  92%  |                                                                              |=================================================================     |  93%  |                                                                              |======================================================================| 100%

  yields <- nassqs_parse(response)

#> Error in nassqs_parse(response): Response is not in the expected json, xml, or csv format. Use `as = 'text'` to see the unparsed response data, or modify your parameters to include `format = 'json'` to make the request in json.

  head(yields)

#> Error in head(yields): object 'yields' not found

  # Get the number of records that would be returned for a given request
  # Equivalent to 'nassqs_record_count(params)'
  response <- nassqs_GET(params, api_path = "get_counts")

#>   |                                                                              |                                                                      |   0%  |                                                                              |=====                                                                 |   7%  |                                                                              |======                                                                |   8%  |                                                                              |==============                                                        |  20%  |                                                                              |===================                                                   |  27%  |                                                                              |================================                                      |  46%  |                                                                              |=============================================                         |  64%  |                                                                              |==========================================================            |  83%  |                                                                              |===========================================================           |  85%  |                                                                              |=============================================================         |  86%  |                                                                              |======================================================================| 100%

  records <- nassqs_parse(response)

#> Error in nassqs_parse(response): Response is not in the expected json, xml, or csv format. Use `as = 'text'` to see the unparsed response data, or modify your parameters to include `format = 'json'` to make the request in json.

  records

#> Error in eval(expr, envir, enclos): object 'records' not found

  # Get the list of allowable values for the parameters 'statisticcat_desc'
  # Equivalent to 'nassqs_param_values("statisticcat_desc")'
  req <- nassqs_GET(list(param = "statisticcat_desc"),
                    api_path = "get_param_values")

#>   |                                                                              |                                                                      |   0%  |                                                                              |=                                                                     |   2%  |                                                                              |===                                                                   |   4%  |                                                                              |===                                                                   |   5%  |                                                                              |=======                                                               |  10%  |                                                                              |==============                                                        |  20%  |                                                                              |=====================                                                 |  31%  |                                                                              |=============================                                         |  42%  |                                                                              |=====================================                                 |  53%  |                                                                              |======================================                                |  54%  |                                                                              |======================================                                |  55%  |                                                                              |==============================================                        |  66%  |                                                                              |===============================================                       |  67%  |                                                                              |===============================================                       |  68%  |                                                                              |=======================================================               |  79%  |                                                                              |========================================================              |  80%  |                                                                              |========================================================              |  81%  |                                                                              |================================================================      |  91%  |                                                                              |=================================================================     |  92%  |                                                                              |=================================================================     |  93%  |                                                                              |======================================================================| 100%

  statisticcat_desc_values <- nassqs_parse(req, as = "list")

#> Error in nassqs_parse(req, as = "list"): Response is not in the expected json, xml, or csv format. Use `as = 'text'` to see the unparsed response data, or modify your parameters to include `format = 'json'` to make the request in json.

  head(statisticcat_desc_values)

#> Error in head(statisticcat_desc_values): object 'statisticcat_desc_values' not found