Issue a GET request to the NASS 'Quick Stats' API

This is the workhorse of the package that provides the core request functionality to the NASS 'Quick Stats' API: https://quickstats.nass.usda.gov/api. In most cases nassqs() or other high-level functions should be used. nassqs_GET() uses httr::GET() to make a HTTP GET request, which returns a request object which must then be parsed to a data.frame, list, or other R object. Higher-level functions will do that parsing automatically. However, if you need access to the request object directly, nassqs_GET() provides that.

nassqs_GET(..., api_path = c("api_GET", "get_param_values",
  "get_counts"))

Arguments

...
either a named list of parameters or a series of parameters to use in the query

api_path
the API path that determines the type of request being made

Value

a httr::GET() response object

Examples

# \donttest{
  # Yields for corn in 2012 in Washington
  params <- list(commodity_name = "CORN",
                 year = 2012,
                 agg_level_desc = "STATE",
                 state_alpha = "WA",
                 statisticcat_desc = "YIELD")

  # Returns a request object that must be parsed either manually or
  # by using nassqs_parse()
  response <- nassqs_GET(params)

#>   |                                                                              |                                                                      |   0%  |                                                                              |==                                                                    |   4%  |                                                                              |===                                                                   |   4%  |                                                                              |========                                                              |  11%  |                                                                              |===============                                                       |  21%  |                                                                              |======================                                                |  31%  |                                                                              |=======================                                               |  32%  |                                                                              |=======================                                               |  33%  |                                                                              |==============================                                        |  43%  |                                                                              |=====================================                                 |  53%  |                                                                              |======================================                                |  54%  |                                                                              |=======================================                               |  55%  |                                                                              |==============================================                        |  66%  |                                                                              |===============================================                       |  66%  |                                                                              |===============================================                       |  67%  |                                                                              |======================================================                |  78%  |                                                                              |=======================================================               |  78%  |                                                                              |========================================================              |  79%  |                                                                              |===============================================================       |  90%  |                                                                              |================================================================      |  91%  |                                                                              |======================================================================| 100%

  yields <- nassqs_parse(response)
  head(yields)

#>   region_desc domaincat_desc state_ansi state_name  country_name
#> 1              NOT SPECIFIED         53 WASHINGTON UNITED STATES
#> 2              NOT SPECIFIED         53 WASHINGTON UNITED STATES
#> 3              NOT SPECIFIED         53 WASHINGTON UNITED STATES
#> 4              NOT SPECIFIED         53 WASHINGTON UNITED STATES
#> 5              NOT SPECIFIED         53 WASHINGTON UNITED STATES
#> 6              NOT SPECIFIED         53 WASHINGTON UNITED STATES
#>          util_practice_desc county_ansi agg_level_desc end_code  group_desc
#> 1 ALL UTILIZATION PRACTICES                      STATE       00 FIELD CROPS
#> 2 ALL UTILIZATION PRACTICES                      STATE       00 FIELD CROPS
#> 3 ALL UTILIZATION PRACTICES                      STATE       00 FIELD CROPS
#> 4 ALL UTILIZATION PRACTICES                      STATE       00 FIELD CROPS
#> 5 ALL UTILIZATION PRACTICES                      STATE       00 FIELD CROPS
#> 6 ALL UTILIZATION PRACTICES                      STATE       00 FIELD CROPS
#>   week_ending freq_desc           load_time zip_5 watershed_desc CV (%)
#> 1                ANNUAL 2012-12-31 00:00:00                         3.0
#> 2                ANNUAL 2012-12-31 00:00:00                         2.5
#> 3                ANNUAL 2012-12-31 00:00:00                         4.6
#> 4                ANNUAL 2012-12-31 00:00:00                         (D)
#> 5                ANNUAL 2012-12-31 00:00:00                         1.0
#> 6                ANNUAL 2012-12-31 00:00:00                         (D)
#>   asd_desc county_code domain_desc county_name source_desc
#> 1                            TOTAL                  CENSUS
#> 2                            TOTAL                  CENSUS
#> 3                            TOTAL                  CENSUS
#> 4                            TOTAL                  CENSUS
#> 5                            TOTAL                  CENSUS
#> 6                            TOTAL                  CENSUS
#>       prodn_practice_desc asd_code congr_district_code year statisticcat_desc
#> 1  IRRIGATED, ENTIRE CROP                              2012             YIELD
#> 2 IRRIGATED, NONE OF CROP                              2012             YIELD
#> 3 IRRIGATED, PART OF CROP                              2012             YIELD
#> 4  IRRIGATED, ENTIRE CROP                              2012             YIELD
#> 5 IRRIGATED, NONE OF CROP                              2012             YIELD
#> 6 IRRIGATED, PART OF CROP                              2012             YIELD
#>                                class_desc location_desc state_alpha
#> 1                             ALL CLASSES    WASHINGTON          WA
#> 2                             ALL CLASSES    WASHINGTON          WA
#> 3                             ALL CLASSES    WASHINGTON          WA
#> 4 DRY EDIBLE, (EXCL LIMA), INCL CHICKPEAS    WASHINGTON          WA
#> 5 DRY EDIBLE, (EXCL LIMA), INCL CHICKPEAS    WASHINGTON          WA
#> 6 DRY EDIBLE, (EXCL LIMA), INCL CHICKPEAS    WASHINGTON          WA
#>   commodity_desc sector_desc                Value country_code state_fips_code
#> 1         BARLEY       CROPS                 86.7         9000              53
#> 2         BARLEY       CROPS                 68.3         9000              53
#> 3         BARLEY       CROPS                 72.6         9000              53
#> 4          BEANS       CROPS                  (D)         9000              53
#> 5          BEANS       CROPS                 17.2         9000              53
#> 6          BEANS       CROPS                  (D)         9000              53
#>   begin_code watershed_code  unit_desc
#> 1         00       00000000  BU / ACRE
#> 2         00       00000000  BU / ACRE
#> 3         00       00000000  BU / ACRE
#> 4         00       00000000 CWT / ACRE
#> 5         00       00000000 CWT / ACRE
#> 6         00       00000000 CWT / ACRE
#>                                                                                                short_desc
#> 1                                           BARLEY, IRRIGATED, ENTIRE CROP - YIELD, MEASURED IN BU / ACRE
#> 2                                          BARLEY, IRRIGATED, NONE OF CROP - YIELD, MEASURED IN BU / ACRE
#> 3                                          BARLEY, IRRIGATED, PART OF CROP - YIELD, MEASURED IN BU / ACRE
#> 4  BEANS, DRY EDIBLE, (EXCL LIMA), INCL CHICKPEAS, IRRIGATED, ENTIRE CROP - YIELD, MEASURED IN CWT / ACRE
#> 5 BEANS, DRY EDIBLE, (EXCL LIMA), INCL CHICKPEAS, IRRIGATED, NONE OF CROP - YIELD, MEASURED IN CWT / ACRE
#> 6 BEANS, DRY EDIBLE, (EXCL LIMA), INCL CHICKPEAS, IRRIGATED, PART OF CROP - YIELD, MEASURED IN CWT / ACRE
#>   reference_period_desc
#> 1                  YEAR
#> 2                  YEAR
#> 3                  YEAR
#> 4                  YEAR
#> 5                  YEAR
#> 6                  YEAR

  # Get the number of records that would be returned for a given request
  # Equivalent to 'nassqs_record_count(params)'
  response <- nassqs_GET(params, api_path = "get_counts")

#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%

  records <- nassqs_parse(response)
  records

#> $count
#> [1] 154
#> 

  # Get the list of allowable values for the parameters 'statisticcat_desc'
  # Equivalent to 'nassqs_param_values("statisticcat_desc")'
  req <- nassqs_GET(list(param = "statisticcat_desc"),
                    api_path = "get_param_values")

#>   |                                                                              |                                                                      |   0%  |                                                                              |====================================                                  |  51%  |                                                                              |========================================================              |  79%  |                                                                              |======================================================================| 100%

  statisticcat_desc_values <- nassqs_parse(req, as = "list")
  head(statisticcat_desc_values)

#> $statisticcat_desc
#>   [1] "ACCESSIBILITY"                                
#>   [2] "ACCESSIBILITY, 5 YEAR AVG"                    
#>   [3] "ACCESSIBILITY, PREVIOUS YEAR"                 
#>   [4] "ACTIVE GINS"                                  
#>   [5] "ACTIVITY"                                     
#>   [6] "ACTIVITY, 5 YEAR AVG"                         
#>   [7] "ACTIVITY, PREVIOUS YEAR"                      
#>   [8] "ADDED & REPLACED"                             
#>   [9] "AGE, AVG"                                     
#>  [10] "APPLICATIONS"                                 
#>  [11] "AREA"                                         
#>  [12] "AREA BEARING"                                 
#>  [13] "AREA BEARING & NON-BEARING"                   
#>  [14] "AREA CERTIFIED"                               
#>  [15] "AREA CROPLAND"                                
#>  [16] "AREA FILLED"                                  
#>  [17] "AREA FLOODED"                                 
#>  [18] "AREA GRAZED"                                  
#>  [19] "AREA GROWN"                                   
#>  [20] "AREA HARVESTED"                               
#>  [21] "AREA IN PRODUCTION"                           
#>  [22] "AREA IN PRODUCTION PREVIOUS CENSUS YEAR"      
#>  [23] "AREA IRRIGATED"                               
#>  [24] "AREA IRRIGATED PREVIOUS CENSUS YEAR"          
#>  [25] "AREA NON-BEARING"                             
#>  [26] "AREA NOT HARVESTED"                           
#>  [27] "AREA OF WATER"                                
#>  [28] "AREA OPERATED"                                
#>  [29] "AREA PLANTED"                                 
#>  [30] "AREA PLANTED, NET"                            
#>  [31] "AREA POLLINATED, PAID BASIS"                  
#>  [32] "AREA RENTED"                                  
#>  [33] "ASSET VALUE"                                  
#>  [34] "AVAILABILITY"                                 
#>  [35] "AVAILABILITY, 5 YEAR AVG"                     
#>  [36] "AVAILABILITY, PREVIOUS YEAR"                  
#>  [37] "BEING MOLTED"                                 
#>  [38] "BLOOM DATE"                                   
#>  [39] "BOLL COUNT"                                   
#>  [40] "BRED"                                         
#>  [41] "BROKEN"                                       
#>  [42] "CAGES"                                        
#>  [43] "CALF CROP"                                    
#>  [44] "CAPACITY"                                     
#>  [45] "CAPTURED"                                     
#>  [46] "CLIPPED"                                      
#>  [47] "CONDEMNED, ANTE-MORTEM"                       
#>  [48] "CONDEMNED, POST-MORTEM"                       
#>  [49] "CONDEMNED, POST-MORTEM, AIRSACCULITIS"        
#>  [50] "CONDEMNED, POST-MORTEM, ALL CAUSES"           
#>  [51] "CONDEMNED, POST-MORTEM, BRUISES"              
#>  [52] "CONDEMNED, POST-MORTEM, CADAVER"              
#>  [53] "CONDEMNED, POST-MORTEM, CONTAMINATION"        
#>  [54] "CONDEMNED, POST-MORTEM, LEUKOSIS"             
#>  [55] "CONDEMNED, POST-MORTEM, OTHER REASONS"        
#>  [56] "CONDEMNED, POST-MORTEM, OVERSCALD"            
#>  [57] "CONDEMNED, POST-MORTEM, SEPTICAEMIA"          
#>  [58] "CONDEMNED, POST-MORTEM, SYNOVITIS"            
#>  [59] "CONDEMNED, POST-MORTEM, TUBERCULOSIS"         
#>  [60] "CONDEMNED, POST-MORTEM, TUMORS"               
#>  [61] "CONDITION"                                    
#>  [62] "CONDITION, 5 YEAR AVG"                        
#>  [63] "CONDITION, PREVIOUS YEAR"                     
#>  [64] "CONDUCT BUSINESS"                             
#>  [65] "COVER"                                        
#>  [66] "COVER, PREVIOUS YEAR"                         
#>  [67] "CRUSHED"                                      
#>  [68] "DAMAGE"                                       
#>  [69] "DAMAGE, 5 YEAR AVG"                           
#>  [70] "DAMAGE, PREVIOUS YEAR"                        
#>  [71] "DAYS SUITABLE"                                
#>  [72] "DAYS SUITABLE, PREVIOUS YEAR"                 
#>  [73] "DEBT"                                         
#>  [74] "DELIVERED VALUE"                              
#>  [75] "DEPTH TO BOWLS"                               
#>  [76] "DEPTH TO WATER"                               
#>  [77] "DEPTH, AVG"                                   
#>  [78] "DEPTH, AVG, PREVIOUS YEAR"                    
#>  [79] "DEVICES"                                      
#>  [80] "DISAPPEARANCE"                                
#>  [81] "DISAPPEARANCE, OTHER"                         
#>  [82] "DISTRIBUTION"                                 
#>  [83] "EAR COUNT"                                    
#>  [84] "EGGS IN INCUBATORS"                           
#>  [85] "EGGS SET"                                     
#>  [86] "END DATE, AVG"                                
#>  [87] "END DATE, MAX"                                
#>  [88] "ENGINE SIZE"                                  
#>  [89] "EXPENSE"                                      
#>  [90] "FACILITIES"                                   
#>  [91] "FARM USE"                                     
#>  [92] "FARROWED"                                     
#>  [93] "FAT TEST"                                     
#>  [94] "FEEDLOTS"                                     
#>  [95] "FRUIT SIZE"                                   
#>  [96] "FRUIT SIZE, 5 YEAR AVG"                       
#>  [97] "FRUIT SIZE, PREVIOUS YEAR"                    
#>  [98] "GAIN"                                         
#>  [99] "GINNED BALES"                                 
#> [100] "GROSS INCOME"                                 
#> [101] "HARVEST LOSS"                                 
#> [102] "HATCHED"                                      
#> [103] "HEAD COUNT"                                   
#> [104] "HEIGHT, AVG"                                  
#> [105] "HEIGHT, AVG, 5 YEAR AVG"                      
#> [106] "HEIGHT, AVG, PREVIOUS YEAR"                   
#> [107] "HOURS OF OPERATION"                           
#> [108] "INDEX FOR PRICE PAID, 1910 - 1914"            
#> [109] "INDEX FOR PRICE PAID, 1990 - 1992"            
#> [110] "INDEX FOR PRICE PAID, 2011"                   
#> [111] "INDEX FOR PRICE RECEIVED, 1910 - 1914"        
#> [112] "INDEX FOR PRICE RECEIVED, 1990 - 1992"        
#> [113] "INDEX FOR PRICE RECEIVED, 2011"               
#> [114] "INFESTATION"                                  
#> [115] "INFESTATION, PREVIOUS YEAR"                   
#> [116] "INVENTORY"                                    
#> [117] "INVENTORY, AVG"                               
#> [118] "INVENTORY, MAX"                               
#> [119] "INVENTORY, PEAK"                              
#> [120] "KID CROP"                                     
#> [121] "LAMB CROP"                                    
#> [122] "LAMBING RATE"                                 
#> [123] "LANDLORDS"                                    
#> [124] "LENGTH OF SEASON, AVG"                        
#> [125] "LITTER RATE"                                  
#> [126] "LOSS"                                         
#> [127] "LOSS, COLONY COLLAPSE DISORDER"               
#> [128] "LOSS, DEADOUT"                                
#> [129] "LOSS, DEATH"                                  
#> [130] "LOSS, DEATH & RENDERED"                       
#> [131] "LOSS, DEATH, 5 YEAR AVG"                      
#> [132] "LOSS, DEATH, PREVIOUS YEAR"                   
#> [133] "MILKFAT"                                      
#> [134] "MILL SALE"                                    
#> [135] "MILLED"                                       
#> [136] "MILLING CAPACITY"                             
#> [137] "MOISTURE"                                     
#> [138] "MOISTURE CONTENT"                             
#> [139] "MOISTURE, 5 YEAR AVG"                         
#> [140] "MOISTURE, PREVIOUS YEAR"                      
#> [141] "MOLT COMPLETED"                               
#> [142] "MOVEMENT"                                     
#> [143] "MOVEMENT, 5 YEAR AVG"                         
#> [144] "MOVEMENT, PREVIOUS YEAR"                      
#> [145] "NET INCOME"                                   
#> [146] "NUT SET"                                      
#> [147] "NUT SET, 5 YEAR AVG"                          
#> [148] "NUT SET, PREVIOUS YEAR"                       
#> [149] "OPERATING PRESSURE"                           
#> [150] "OPERATIONS"                                   
#> [151] "OPERATORS"                                    
#> [152] "OTHER SALE"                                   
#> [153] "PARTNERS, AVG"                                
#> [154] "PENS"                                         
#> [155] "PERSONS IN HOUSEHOLD"                         
#> [156] "PEST MGMT"                                    
#> [157] "PIG CROP"                                     
#> [158] "PLACEMENTS"                                   
#> [159] "PLACEMENTS, INTENDED"                         
#> [160] "PLACEMENTS, NET"                              
#> [161] "PLANT POPULATION"                             
#> [162] "POD COUNT"                                    
#> [163] "POLLINATION"                                  
#> [164] "PONDS"                                        
#> [165] "PRICE PAID"                                   
#> [166] "PRICE REACTION"                               
#> [167] "PRICE REACTION, DECREASE"                     
#> [168] "PRICE REACTION, INCREASE"                     
#> [169] "PRICE REACTION, NO CHANGE"                    
#> [170] "PRICE RECEIVED"                               
#> [171] "PRICE RECEIVED AFTER REPORT"                  
#> [172] "PRICE RECEIVED PRIOR TO CLOSING"              
#> [173] "PRICE RECEIVED, 3 YEAR AVG"                   
#> [174] "PRICE RECEIVED, ADJUSTED BASE"                
#> [175] "PRICE RECEIVED, PARITY"                       
#> [176] "PRICE RECEIVED, PREMIUM"                      
#> [177] "PRICE RECEIVED, SEASONALLY ADJUSTED"          
#> [178] "PROCESSED IN LAB"                             
#> [179] "PRODUCERS"                                    
#> [180] "PRODUCTION"                                   
#> [181] "PROGRESS"                                     
#> [182] "PROGRESS, 5 YEAR AVG"                         
#> [183] "PROGRESS, PREVIOUS YEAR"                      
#> [184] "PUMPING CAPACITY"                             
#> [185] "PUMPING LIFT"                                 
#> [186] "PUMPS"                                        
#> [187] "RACEWAYS"                                     
#> [188] "RATE OF LAY"                                  
#> [189] "RATIO"                                        
#> [190] "RECEIPTS"                                     
#> [191] "RECEIVING"                                    
#> [192] "RECEIVING, 5 YEAR AVG"                        
#> [193] "RECEIVING, PREVIOUS YEAR"                     
#> [194] "RELATIVE WEIGHT"                              
#> [195] "RELATIVE WEIGHT, 5 YEAR MOVING AVERAGE"       
#> [196] "REMOVAL FOR PROCESSING"                       
#> [197] "REMOVAL FOR PROCESSING, EDIBLE USE"           
#> [198] "REMOVAL FOR PROCESSING, INEDIBLE USE"         
#> [199] "REPAYMENTS"                                   
#> [200] "ROW WIDTH"                                    
#> [201] "SALES"                                        
#> [202] "SALES & DISTRIBUTION"                         
#> [203] "SALES FOR SLAUGHTER"                          
#> [204] "SALES IN CONVENTIONAL MARKETS"                
#> [205] "SALES IN ORGANIC MARKETS"                     
#> [206] "SAMPLES"                                      
#> [207] "SAVINGS"                                      
#> [208] "SEASONAL ADJUSTMENT FACTOR FOR PRICE RECEIVED"
#> [209] "SEED FOR PLANTING"                            
#> [210] "SHIPMENTS"                                    
#> [211] "SHIPMENTS IN"                                 
#> [212] "SHORN"                                        
#> [213] "SHRINK"                                       
#> [214] "SLAUGHTERED"                                  
#> [215] "SOURCE"                                       
#> [216] "SOURCE, 5 YEAR AVG"                           
#> [217] "SOURCE, PREVIOUS YEAR"                        
#> [218] "SPINDLES"                                     
#> [219] "START DATE"                                   
#> [220] "START DATE, 5 YEAR AVG"                       
#> [221] "START DATE, AVG"                              
#> [222] "START DATE, MIN"                              
#> [223] "START DATE, PREVIOUS YEAR"                    
#> [224] "STOCKS"                                       
#> [225] "STOCKS, CURRENT YEAR"                         
#> [226] "STOCKS, PREVIOUS YEAR"                        
#> [227] "SUCROSE"                                      
#> [228] "SUPPLIES"                                     
#> [229] "SUPPLIES, 5 YEAR AVG"                         
#> [230] "SUPPLIES, PREVIOUS YEAR"                      
#> [231] "TANKS"                                        
#> [232] "TAPS"                                         
#> [233] "TENANTS"                                      
#> [234] "TIME OPERATED"                                
#> [235] "TIME WORKED"                                  
#> [236] "TOTAL AREA"                                   
#> [237] "TREATED"                                      
#> [238] "USAGE"                                        
#> [239] "VALUE"                                        
#> [240] "VALUE OF POLLINATION"                         
#> [241] "VOLUME"                                       
#> [242] "VOLUME HANDLED"                               
#> [243] "WAGE RATE"                                    
#> [244] "WAGE RATE, BASE"                              
#> [245] "WATER APPLIED"                                
#> [246] "WATER RECEIVED"                               
#> [247] "WELL DEPTH"                                   
#> [248] "WELLS"                                        
#> [249] "WORKERS"                                      
#> [250] "WORKERS, AVG"                                 
#> [251] "WORKERS, PEAK"                                
#> [252] "YEARS ON ANY OPERATION, AVG"                  
#> [253] "YEARS ON PRESENT OPERATION, AVG"              
#> [254] "YEARS RENTED TO TENANT, AVG"                  
#> [255] "YIELD"                                        
#> 

# }