Issue a GET request to the NASS 'Quick Stats' API

This is the workhorse of the package that provides the core request functionality to the NASS 'Quick Stats' API: https://quickstats.nass.usda.gov/api. In most cases nassqs() or other high-level functions should be used. nassqs_GET uses httr::GET() to make a HTTP GET request, which returns a request object which must then be parsed to a data.frame, list, or other R object. Higher-level functions will do that parsing automatically. However, if you need access to the request object directly, nassqs_GET provides that.

nassqs_GET(..., api_path = c("api_GET", "get_param_values",
  "get_counts"))

Arguments

...
either a named list of parameters or a series of parameters to use in the query

api_path
the API path that determines the type of request being made

Value

a httr::GET() response object

Examples

  # Yields for corn in 2012 in Washington
  params <- list(commodity_name = "CORN",
                 year = 2012,
                 agg_level_desc = "STATE",
                 state_alpha = "WA",
                 statisticcat_desc = "YIELD")

  # Returns a request object that must be parsed either manually or
  # by using nassqs_parse()
  response <- nassqs_GET(params)
  yields <- nassqs_parse(response)
  head(yields)

#>   source_desc asd_desc commodity_desc watershed_code        util_practice_desc
#> 1      CENSUS                  BARLEY       00000000 ALL UTILIZATION PRACTICES
#> 2      CENSUS                  BARLEY       00000000 ALL UTILIZATION PRACTICES
#> 3      CENSUS                  BARLEY       00000000 ALL UTILIZATION PRACTICES
#> 4      CENSUS                   BEANS       00000000 ALL UTILIZATION PRACTICES
#> 5      CENSUS                   BEANS       00000000 ALL UTILIZATION PRACTICES
#> 6      CENSUS                   BEANS       00000000 ALL UTILIZATION PRACTICES
#>   reference_period_desc  unit_desc state_ansi agg_level_desc
#> 1                  YEAR  BU / ACRE         53          STATE
#> 2                  YEAR  BU / ACRE         53          STATE
#> 3                  YEAR  BU / ACRE         53          STATE
#> 4                  YEAR CWT / ACRE         53          STATE
#> 5                  YEAR CWT / ACRE         53          STATE
#> 6                  YEAR CWT / ACRE         53          STATE
#>             load_time state_fips_code sector_desc state_name
#> 1 2012-12-31 00:00:00              53       CROPS WASHINGTON
#> 2 2012-12-31 00:00:00              53       CROPS WASHINGTON
#> 3 2012-12-31 00:00:00              53       CROPS WASHINGTON
#> 4 2012-12-31 00:00:00              53       CROPS WASHINGTON
#> 5 2012-12-31 00:00:00              53       CROPS WASHINGTON
#> 6 2012-12-31 00:00:00              53       CROPS WASHINGTON
#>   congr_district_code  country_name county_name
#> 1                     UNITED STATES            
#> 2                     UNITED STATES            
#> 3                     UNITED STATES            
#> 4                     UNITED STATES            
#> 5                     UNITED STATES            
#> 6                     UNITED STATES            
#>                                                                                                short_desc
#> 1                                           BARLEY, IRRIGATED, ENTIRE CROP - YIELD, MEASURED IN BU / ACRE
#> 2                                          BARLEY, IRRIGATED, NONE OF CROP - YIELD, MEASURED IN BU / ACRE
#> 3                                          BARLEY, IRRIGATED, PART OF CROP - YIELD, MEASURED IN BU / ACRE
#> 4  BEANS, DRY EDIBLE, (EXCL LIMA), INCL CHICKPEAS, IRRIGATED, ENTIRE CROP - YIELD, MEASURED IN CWT / ACRE
#> 5 BEANS, DRY EDIBLE, (EXCL LIMA), INCL CHICKPEAS, IRRIGATED, NONE OF CROP - YIELD, MEASURED IN CWT / ACRE
#> 6 BEANS, DRY EDIBLE, (EXCL LIMA), INCL CHICKPEAS, IRRIGATED, PART OF CROP - YIELD, MEASURED IN CWT / ACRE
#>   region_desc domain_desc domaincat_desc                Value zip_5  group_desc
#> 1                   TOTAL  NOT SPECIFIED                 86.7       FIELD CROPS
#> 2                   TOTAL  NOT SPECIFIED                 68.3       FIELD CROPS
#> 3                   TOTAL  NOT SPECIFIED                 72.6       FIELD CROPS
#> 4                   TOTAL  NOT SPECIFIED                  (D)       FIELD CROPS
#> 5                   TOTAL  NOT SPECIFIED                 17.2       FIELD CROPS
#> 6                   TOTAL  NOT SPECIFIED                  (D)       FIELD CROPS
#>   county_ansi CV (%) week_ending                              class_desc
#> 1                3.0                                         ALL CLASSES
#> 2                2.5                                         ALL CLASSES
#> 3                4.6                                         ALL CLASSES
#> 4                (D)             DRY EDIBLE, (EXCL LIMA), INCL CHICKPEAS
#> 5                1.0             DRY EDIBLE, (EXCL LIMA), INCL CHICKPEAS
#> 6                (D)             DRY EDIBLE, (EXCL LIMA), INCL CHICKPEAS
#>   location_desc statisticcat_desc freq_desc county_code begin_code asd_code
#> 1    WASHINGTON             YIELD    ANNUAL                     00         
#> 2    WASHINGTON             YIELD    ANNUAL                     00         
#> 3    WASHINGTON             YIELD    ANNUAL                     00         
#> 4    WASHINGTON             YIELD    ANNUAL                     00         
#> 5    WASHINGTON             YIELD    ANNUAL                     00         
#> 6    WASHINGTON             YIELD    ANNUAL                     00         
#>   watershed_desc state_alpha     prodn_practice_desc year country_code end_code
#> 1                         WA  IRRIGATED, ENTIRE CROP 2012         9000       00
#> 2                         WA IRRIGATED, NONE OF CROP 2012         9000       00
#> 3                         WA IRRIGATED, PART OF CROP 2012         9000       00
#> 4                         WA  IRRIGATED, ENTIRE CROP 2012         9000       00
#> 5                         WA IRRIGATED, NONE OF CROP 2012         9000       00
#> 6                         WA IRRIGATED, PART OF CROP 2012         9000       00

  # Get the number of records that would be returned for a given request
  # Equivalent to 'nassqs_record_count(params)'
  response <- nassqs_GET(params, api_path = "get_counts")
  records <- nassqs_parse(response)
  records

#> $count
#> [1] 154
#> 

  # Get the list of allowable values for the parameters 'statisticcat_desc'
  # Equivalent to 'nassqs_param_values("statisticcat_desc")'
  req <- nassqs_GET(list(param = "statisticcat_desc"),
                    api_path = "get_param_values")
  statisticcat_desc_values <- nassqs_parse(req, as = "list")
  head(statisticcat_desc_values)

#> $statisticcat_desc
#>   [1] "ACCESSIBILITY"                                
#>   [2] "ACCESSIBILITY, 5 YEAR AVG"                    
#>   [3] "ACCESSIBILITY, PREVIOUS YEAR"                 
#>   [4] "ACTIVE GINS"                                  
#>   [5] "ACTIVITY"                                     
#>   [6] "ACTIVITY, 5 YEAR AVG"                         
#>   [7] "ACTIVITY, PREVIOUS YEAR"                      
#>   [8] "ADDED & REPLACED"                             
#>   [9] "AGE, AVG"                                     
#>  [10] "APPLICATIONS"                                 
#>  [11] "AREA"                                         
#>  [12] "AREA BEARING"                                 
#>  [13] "AREA BEARING & NON-BEARING"                   
#>  [14] "AREA CERTIFIED"                               
#>  [15] "AREA CROPLAND"                                
#>  [16] "AREA FILLED"                                  
#>  [17] "AREA FLOODED"                                 
#>  [18] "AREA GRAZED"                                  
#>  [19] "AREA GROWN"                                   
#>  [20] "AREA HARVESTED"                               
#>  [21] "AREA IN PRODUCTION"                           
#>  [22] "AREA IN PRODUCTION PREVIOUS CENSUS YEAR"      
#>  [23] "AREA IRRIGATED"                               
#>  [24] "AREA IRRIGATED PREVIOUS CENSUS YEAR"          
#>  [25] "AREA NON-BEARING"                             
#>  [26] "AREA NOT HARVESTED"                           
#>  [27] "AREA OF WATER"                                
#>  [28] "AREA OPERATED"                                
#>  [29] "AREA PLANTED"                                 
#>  [30] "AREA PLANTED, NET"                            
#>  [31] "AREA POLLINATED, PAID BASIS"                  
#>  [32] "AREA RENTED"                                  
#>  [33] "ASSET VALUE"                                  
#>  [34] "AVAILABILITY"                                 
#>  [35] "AVAILABILITY, 5 YEAR AVG"                     
#>  [36] "AVAILABILITY, PREVIOUS YEAR"                  
#>  [37] "BEING MOLTED"                                 
#>  [38] "BLOOM DATE"                                   
#>  [39] "BOLL COUNT"                                   
#>  [40] "BRED"                                         
#>  [41] "BROKEN"                                       
#>  [42] "CAGES"                                        
#>  [43] "CALF CROP"                                    
#>  [44] "CAPACITY"                                     
#>  [45] "CAPTURED"                                     
#>  [46] "CLIPPED"                                      
#>  [47] "CONDEMNED, ANTE-MORTEM"                       
#>  [48] "CONDEMNED, POST-MORTEM"                       
#>  [49] "CONDEMNED, POST-MORTEM, AIRSACCULITIS"        
#>  [50] "CONDEMNED, POST-MORTEM, ALL CAUSES"           
#>  [51] "CONDEMNED, POST-MORTEM, BRUISES"              
#>  [52] "CONDEMNED, POST-MORTEM, CADAVER"              
#>  [53] "CONDEMNED, POST-MORTEM, CONTAMINATION"        
#>  [54] "CONDEMNED, POST-MORTEM, LEUKOSIS"             
#>  [55] "CONDEMNED, POST-MORTEM, OTHER REASONS"        
#>  [56] "CONDEMNED, POST-MORTEM, OVERSCALD"            
#>  [57] "CONDEMNED, POST-MORTEM, SEPTICAEMIA"          
#>  [58] "CONDEMNED, POST-MORTEM, SYNOVITIS"            
#>  [59] "CONDEMNED, POST-MORTEM, TUBERCULOSIS"         
#>  [60] "CONDEMNED, POST-MORTEM, TUMORS"               
#>  [61] "CONDITION"                                    
#>  [62] "CONDITION, 5 YEAR AVG"                        
#>  [63] "CONDITION, PREVIOUS YEAR"                     
#>  [64] "CONDUCT BUSINESS"                             
#>  [65] "COVER"                                        
#>  [66] "COVER, PREVIOUS YEAR"                         
#>  [67] "CRUSHED"                                      
#>  [68] "DAMAGE"                                       
#>  [69] "DAMAGE, 5 YEAR AVG"                           
#>  [70] "DAMAGE, PREVIOUS YEAR"                        
#>  [71] "DAYS SUITABLE"                                
#>  [72] "DAYS SUITABLE, PREVIOUS YEAR"                 
#>  [73] "DEBT"                                         
#>  [74] "DELIVERED VALUE"                              
#>  [75] "DEPTH TO BOWLS"                               
#>  [76] "DEPTH TO WATER"                               
#>  [77] "DEPTH, AVG"                                   
#>  [78] "DEPTH, AVG, PREVIOUS YEAR"                    
#>  [79] "DEVICES"                                      
#>  [80] "DISAPPEARANCE"                                
#>  [81] "DISAPPEARANCE, OTHER"                         
#>  [82] "DISTRIBUTION"                                 
#>  [83] "EAR COUNT"                                    
#>  [84] "EGGS IN INCUBATORS"                           
#>  [85] "EGGS SET"                                     
#>  [86] "END DATE, AVG"                                
#>  [87] "END DATE, MAX"                                
#>  [88] "ENGINE SIZE"                                  
#>  [89] "EXPENSE"                                      
#>  [90] "FACILITIES"                                   
#>  [91] "FARM USE"                                     
#>  [92] "FARROWED"                                     
#>  [93] "FAT TEST"                                     
#>  [94] "FEEDLOTS"                                     
#>  [95] "FRUIT SIZE"                                   
#>  [96] "FRUIT SIZE, 5 YEAR AVG"                       
#>  [97] "FRUIT SIZE, PREVIOUS YEAR"                    
#>  [98] "GAIN"                                         
#>  [99] "GINNED BALES"                                 
#> [100] "GROSS INCOME"                                 
#> [101] "HARVEST LOSS"                                 
#> [102] "HATCHED"                                      
#> [103] "HEAD COUNT"                                   
#> [104] "HEIGHT, AVG"                                  
#> [105] "HEIGHT, AVG, 5 YEAR AVG"                      
#> [106] "HEIGHT, AVG, PREVIOUS YEAR"                   
#> [107] "HOURS OF OPERATION"                           
#> [108] "INDEX FOR PRICE PAID, 1910 - 1914"            
#> [109] "INDEX FOR PRICE PAID, 1990 - 1992"            
#> [110] "INDEX FOR PRICE PAID, 2011"                   
#> [111] "INDEX FOR PRICE RECEIVED, 1910 - 1914"        
#> [112] "INDEX FOR PRICE RECEIVED, 1990 - 1992"        
#> [113] "INDEX FOR PRICE RECEIVED, 2011"               
#> [114] "INFESTATION"                                  
#> [115] "INFESTATION, PREVIOUS YEAR"                   
#> [116] "INVENTORY"                                    
#> [117] "INVENTORY, AVG"                               
#> [118] "INVENTORY, MAX"                               
#> [119] "INVENTORY, PEAK"                              
#> [120] "KID CROP"                                     
#> [121] "LAMB CROP"                                    
#> [122] "LAMBING RATE"                                 
#> [123] "LANDLORDS"                                    
#> [124] "LENGTH OF SEASON, AVG"                        
#> [125] "LITTER RATE"                                  
#> [126] "LOSS"                                         
#> [127] "LOSS, COLONY COLLAPSE DISORDER"               
#> [128] "LOSS, DEADOUT"                                
#> [129] "LOSS, DEATH"                                  
#> [130] "LOSS, DEATH & RENDERED"                       
#> [131] "LOSS, DEATH, 5 YEAR AVG"                      
#> [132] "LOSS, DEATH, PREVIOUS YEAR"                   
#> [133] "MILKFAT"                                      
#> [134] "MILL SALE"                                    
#> [135] "MILLED"                                       
#> [136] "MILLING CAPACITY"                             
#> [137] "MOISTURE"                                     
#> [138] "MOISTURE CONTENT"                             
#> [139] "MOISTURE, 5 YEAR AVG"                         
#> [140] "MOISTURE, PREVIOUS YEAR"                      
#> [141] "MOLT COMPLETED"                               
#> [142] "MOVEMENT"                                     
#> [143] "MOVEMENT, 5 YEAR AVG"                         
#> [144] "MOVEMENT, PREVIOUS YEAR"                      
#> [145] "NET INCOME"                                   
#> [146] "NUT SET"                                      
#> [147] "NUT SET, 5 YEAR AVG"                          
#> [148] "NUT SET, PREVIOUS YEAR"                       
#> [149] "OPERATING PRESSURE"                           
#> [150] "OPERATIONS"                                   
#> [151] "OPERATORS"                                    
#> [152] "OTHER SALE"                                   
#> [153] "PARTNERS, AVG"                                
#> [154] "PENS"                                         
#> [155] "PERSONS IN HOUSEHOLD"                         
#> [156] "PEST MGMT"                                    
#> [157] "PIG CROP"                                     
#> [158] "PLACEMENTS"                                   
#> [159] "PLACEMENTS, INTENDED"                         
#> [160] "PLACEMENTS, NET"                              
#> [161] "PLANT POPULATION"                             
#> [162] "POD COUNT"                                    
#> [163] "POLLINATION"                                  
#> [164] "PONDS"                                        
#> [165] "PRICE PAID"                                   
#> [166] "PRICE REACTION"                               
#> [167] "PRICE REACTION, DECREASE"                     
#> [168] "PRICE REACTION, INCREASE"                     
#> [169] "PRICE REACTION, NO CHANGE"                    
#> [170] "PRICE RECEIVED"                               
#> [171] "PRICE RECEIVED AFTER REPORT"                  
#> [172] "PRICE RECEIVED PRIOR TO CLOSING"              
#> [173] "PRICE RECEIVED, 3 YEAR AVG"                   
#> [174] "PRICE RECEIVED, ADJUSTED BASE"                
#> [175] "PRICE RECEIVED, PARITY"                       
#> [176] "PRICE RECEIVED, PREMIUM"                      
#> [177] "PRICE RECEIVED, SEASONALLY ADJUSTED"          
#> [178] "PROCESSED IN LAB"                             
#> [179] "PRODUCERS"                                    
#> [180] "PRODUCTION"                                   
#> [181] "PROGRESS"                                     
#> [182] "PROGRESS, 5 YEAR AVG"                         
#> [183] "PROGRESS, PREVIOUS YEAR"                      
#> [184] "PUMPING CAPACITY"                             
#> [185] "PUMPING LIFT"                                 
#> [186] "PUMPS"                                        
#> [187] "RACEWAYS"                                     
#> [188] "RATE OF LAY"                                  
#> [189] "RATIO"                                        
#> [190] "RECEIPTS"                                     
#> [191] "RECEIVING"                                    
#> [192] "RECEIVING, 5 YEAR AVG"                        
#> [193] "RECEIVING, PREVIOUS YEAR"                     
#> [194] "RELATIVE WEIGHT"                              
#> [195] "RELATIVE WEIGHT, 5 YEAR MOVING AVERAGE"       
#> [196] "REMOVAL FOR PROCESSING"                       
#> [197] "REMOVAL FOR PROCESSING, EDIBLE USE"           
#> [198] "REMOVAL FOR PROCESSING, INEDIBLE USE"         
#> [199] "REPAYMENTS"                                   
#> [200] "ROW WIDTH"                                    
#> [201] "SALES"                                        
#> [202] "SALES & DISTRIBUTION"                         
#> [203] "SALES FOR SLAUGHTER"                          
#> [204] "SALES IN CONVENTIONAL MARKETS"                
#> [205] "SALES IN ORGANIC MARKETS"                     
#> [206] "SAMPLES"                                      
#> [207] "SAVINGS"                                      
#> [208] "SEASONAL ADJUSTMENT FACTOR FOR PRICE RECEIVED"
#> [209] "SEED FOR PLANTING"                            
#> [210] "SHIPMENTS"                                    
#> [211] "SHIPMENTS IN"                                 
#> [212] "SHORN"                                        
#> [213] "SHRINK"                                       
#> [214] "SLAUGHTERED"                                  
#> [215] "SOURCE"                                       
#> [216] "SOURCE, PREVIOUS YEAR"                        
#> [217] "SPINDLES"                                     
#> [218] "START DATE"                                   
#> [219] "START DATE, 5 YEAR AVG"                       
#> [220] "START DATE, AVG"                              
#> [221] "START DATE, MIN"                              
#> [222] "START DATE, PREVIOUS YEAR"                    
#> [223] "STOCKS"                                       
#> [224] "STOCKS, CURRENT YEAR"                         
#> [225] "STOCKS, PREVIOUS YEAR"                        
#> [226] "SUCROSE"                                      
#> [227] "SUPPLIES"                                     
#> [228] "SUPPLIES, 5 YEAR AVG"                         
#> [229] "SUPPLIES, PREVIOUS YEAR"                      
#> [230] "TANKS"                                        
#> [231] "TAPS"                                         
#> [232] "TENANTS"                                      
#> [233] "TIME OPERATED"                                
#> [234] "TIME WORKED"                                  
#> [235] "TOTAL AREA"                                   
#> [236] "TREATED"                                      
#> [237] "USAGE"                                        
#> [238] "VALUE"                                        
#> [239] "VALUE OF POLLINATION"                         
#> [240] "VOLUME"                                       
#> [241] "VOLUME HANDLED"                               
#> [242] "WAGE RATE"                                    
#> [243] "WATER APPLIED"                                
#> [244] "WATER RECEIVED"                               
#> [245] "WELL DEPTH"                                   
#> [246] "WELLS"                                        
#> [247] "WORKERS"                                      
#> [248] "WORKERS, AVG"                                 
#> [249] "WORKERS, PEAK"                                
#> [250] "YEARS ON ANY OPERATION, AVG"                  
#> [251] "YEARS ON PRESENT OPERATION, AVG"              
#> [252] "YEARS RENTED TO TENANT, AVG"                  
#> [253] "YIELD"                                        
#>