Return list of NASS QS parameters.

Contains a simple hard-coded list of all available parameters. If no parameter name is provided, returns a list of all parameters. More information can be found in the API documentation on parameters found at https://quickstats.nass.usda.gov/api#param_define.

nassqs_params(...)

Arguments

...
a parameter, series of parameters, or a list of parameters that you would like a description of. If missing, a list of all available parameters is returned.

Value

a list of all available parameters or a description of a subset

Examples

# Get a list of all available parameters
nassqs_params()

#>  [1] "agg_level_desc"        "asd_code"              "asd_desc"             
#>  [4] "begin_code"            "class_desc"            "commodity_desc"       
#>  [7] "congr_district_code"   "country_code"          "country_name"         
#> [10] "county_ansi"           "county_code"           "county_name"          
#> [13] "CV"                    "domaincat_desc"        "domain_desc"          
#> [16] "end_code"              "freq_desc"             "group_desc"           
#> [19] "load_time"             "location_desc"         "prodn_practice_desc"  
#> [22] "reference_period_desc" "region_desc"           "sector_desc"          
#> [25] "short_desc"            "state_alpha"           "state_ansi"           
#> [28] "state_name"            "state_fips_code"       "statisticcat_desc"    
#> [31] "source_desc"           "unit_desc"             "util_practice_desc"   
#> [34] "Value"                 "watershed_code"        "watershed_desc"       
#> [37] "week_ending"           "year"                  "zip_5"                
#> [40] "format"               

# Get information about specific parameters
nassqs_params("source_desc", "group_desc")

#> [1] "source_desc: Data source. Either 'CENSUS' or 'SURVEY'"       
#> [2] "group_desc: Commodity group category. A subset of the sector"